move metrics to katex-fonts submodule (#1134)
* move metrics to katex-fonts submodule

* update katex-fonts submodule

* update katex-fonts submodule after pulling down changes
