--- src/fontMetricsData.js
@@ -1,2116 +1,0 @@
-export default {
-    "AMS-Regular": {
-        "65": [0, 0.68889, 0, 0, 0.72222],
-        "66": [0, 0.68889, 0, 0, 0.66667],
-        "67": [0, 0.68889, 0, 0, 0.72222],
-        "68": [0, 0.68889, 0, 0, 0.72222],
-        "69": [0, 0.68889, 0, 0, 0.66667],
-        "70": [0, 0.68889, 0, 0, 0.61111],
-        "71": [0, 0.68889, 0, 0, 0.77778],
-        "72": [0, 0.68889, 0, 0, 0.77778],
-        "73": [0, 0.68889, 0, 0, 0.38889],
-        "74": [0.16667, 0.68889, 0, 0, 0.5],
-        "75": [0, 0.68889, 0, 0, 0.77778],
-        "76": [0, 0.68889, 0, 0, 0.66667],
-        "77": [0, 0.68889, 0, 0, 0.94445],
-        "78": [0, 0.68889, 0, 0, 0.72222],
-        "79": [0.16667, 0.68889, 0, 0, 0.77778],
-        "80": [0, 0.68889, 0, 0, 0.61111],
-        "81": [0.16667, 0.68889, 0, 0, 0.77778],
-        "82": [0, 0.68889, 0, 0, 0.72222],
-        "83": [0, 0.68889, 0, 0, 0.55556],
-        "84": [0, 0.68889, 0, 0, 0.66667],
-        "85": [0, 0.68889, 0, 0, 0.72222],
-        "86": [0, 0.68889, 0, 0, 0.72222],
-        "87": [0, 0.68889, 0, 0, 1.0],
-        "88": [0, 0.68889, 0, 0, 0.72222],
-        "89": [0, 0.68889, 0, 0, 0.72222],
-        "90": [0, 0.68889, 0, 0, 0.66667],
-        "107": [0, 0.68889, 0, 0, 0.55556],
-        "165": [0, 0.675, 0.025, 0, 0.75],
-        "174": [0.15559, 0.69224, 0, 0, 0.94666],
-        "240": [0, 0.68889, 0, 0, 0.55556],
-        "295": [0, 0.68889, 0, 0, 0.54028],
-        "710": [0, 0.825, 0, 0, 2.33334],
-        "732": [0, 0.9, 0, 0, 2.33334],
-        "770": [0, 0.825, 0, 0, 2.33334],
-        "771": [0, 0.9, 0, 0, 2.33334],
-        "989": [0.08167, 0.58167, 0, 0, 0.77778],
-        "1008": [0, 0.43056, 0.04028, 0, 0.66667],
-        "8245": [0, 0.54986, 0, 0, 0.275],
-        "8463": [0, 0.68889, 0, 0, 0.54028],
-        "8487": [0, 0.68889, 0, 0, 0.72222],
-        "8498": [0, 0.68889, 0, 0, 0.55556],
-        "8502": [0, 0.68889, 0, 0, 0.66667],
-        "8503": [0, 0.68889, 0, 0, 0.44445],
-        "8504": [0, 0.68889, 0, 0, 0.66667],
-        "8513": [0, 0.68889, 0, 0, 0.63889],
-        "8592": [-0.03598, 0.46402, 0, 0, 0.5],
-        "8594": [-0.03598, 0.46402, 0, 0, 0.5],
-        "8602": [-0.13313, 0.36687, 0, 0, 1.0],
-        "8603": [-0.13313, 0.36687, 0, 0, 1.0],
-        "8606": [0.01354, 0.52239, 0, 0, 1.0],
-        "8608": [0.01354, 0.52239, 0, 0, 1.0],
-        "8610": [0.01354, 0.52239, 0, 0, 1.11111],
-        "8611": [0.01354, 0.52239, 0, 0, 1.11111],
-        "8619": [0, 0.54986, 0, 0, 1.0],
-        "8620": [0, 0.54986, 0, 0, 1.0],
-        "8621": [-0.13313, 0.37788, 0, 0, 1.38889],
-        "8622": [-0.13313, 0.36687, 0, 0, 1.0],
-        "8624": [0, 0.69224, 0, 0, 0.5],
-        "8625": [0, 0.69224, 0, 0, 0.5],
-        "8630": [0, 0.43056, 0, 0, 1.0],
-        "8631": [0, 0.43056, 0, 0, 1.0],
-        "8634": [0.08198, 0.58198, 0, 0, 0.77778],
-        "8635": [0.08198, 0.58198, 0, 0, 0.77778],
-        "8638": [0.19444, 0.69224, 0, 0, 0.41667],
-        "8639": [0.19444, 0.69224, 0, 0, 0.41667],
-        "8642": [0.19444, 0.69224, 0, 0, 0.41667],
-        "8643": [0.19444, 0.69224, 0, 0, 0.41667],
-        "8644": [0.1808, 0.675, 0, 0, 1.0],
-        "8646": [0.1808, 0.675, 0, 0, 1.0],
-        "8647": [0.1808, 0.675, 0, 0, 1.0],
-        "8648": [0.19444, 0.69224, 0, 0, 0.83334],
-        "8649": [0.1808, 0.675, 0, 0, 1.0],
-        "8650": [0.19444, 0.69224, 0, 0, 0.83334],
-        "8651": [0.01354, 0.52239, 0, 0, 1.0],
-        "8652": [0.01354, 0.52239, 0, 0, 1.0],
-        "8653": [-0.13313, 0.36687, 0, 0, 1.0],
-        "8654": [-0.13313, 0.36687, 0, 0, 1.0],
-        "8655": [-0.13313, 0.36687, 0, 0, 1.0],
-        "8666": [0.13667, 0.63667, 0, 0, 1.0],
-        "8667": [0.13667, 0.63667, 0, 0, 1.0],
-        "8669": [-0.13313, 0.37788, 0, 0, 1.0],
-        "8672": [-0.064, 0.437, 0, 0, 1187],
-        "8674": [-0.064, 0.437, 0, 0, 1167],
-        "8705": [0, 0.825, 0, 0, 0.5],
-        "8708": [0, 0.68889, 0, 0, 0.55556],
-        "8709": [0.08167, 0.58167, 0, 0, 0.77778],
-        "8717": [0, 0.43056, 0, 0, 0.42917],
-        "8722": [-0.03598, 0.46402, 0, 0, 0.5],
-        "8724": [0.08198, 0.69224, 0, 0, 0.77778],
-        "8726": [0.08167, 0.58167, 0, 0, 0.77778],
-        "8733": [0, 0.69224, 0, 0, 0.77778],
-        "8736": [0, 0.69224, 0, 0, 0.72222],
-        "8737": [0, 0.69224, 0, 0, 0.72222],
-        "8738": [0.03517, 0.52239, 0, 0, 0.72222],
-        "8739": [0.08167, 0.58167, 0, 0, 0.22222],
-        "8740": [0.25142, 0.74111, 0, 0, 0.27778],
-        "8741": [0.08167, 0.58167, 0, 0, 0.38889],
-        "8742": [0.25142, 0.74111, 0, 0, 0.5],
-        "8756": [0, 0.69224, 0, 0, 0.66667],
-        "8757": [0, 0.69224, 0, 0, 0.66667],
-        "8764": [-0.13313, 0.36687, 0, 0, 0.77778],
-        "8765": [-0.13313, 0.37788, 0, 0, 0.77778],
-        "8769": [-0.13313, 0.36687, 0, 0, 0.77778],
-        "8770": [-0.03625, 0.46375, 0, 0, 0.77778],
-        "8774": [0.30274, 0.79383, 0, 0, 0.77778],
-        "8776": [-0.01688, 0.48312, 0, 0, 0.77778],
-        "8778": [0.08167, 0.58167, 0, 0, 0.77778],
-        "8782": [0.06062, 0.54986, 0, 0, 0.77778],
-        "8783": [0.06062, 0.54986, 0, 0, 0.77778],
-        "8785": [0.08198, 0.58198, 0, 0, 0.77778],
-        "8786": [0.08198, 0.58198, 0, 0, 0.77778],
-        "8787": [0.08198, 0.58198, 0, 0, 0.77778],
-        "8790": [0, 0.69224, 0, 0, 0.77778],
-        "8791": [0.22958, 0.72958, 0, 0, 0.77778],
-        "8796": [0.08198, 0.91667, 0, 0, 0.77778],
-        "8806": [0.25583, 0.75583, 0, 0, 0.77778],
-        "8807": [0.25583, 0.75583, 0, 0, 0.77778],
-        "8808": [0.25142, 0.75726, 0, 0, 0.77778],
-        "8809": [0.25142, 0.75726, 0, 0, 0.77778],
-        "8812": [0.25583, 0.75583, 0, 0, 0.5],
-        "8814": [0.20576, 0.70576, 0, 0, 0.77778],
-        "8815": [0.20576, 0.70576, 0, 0, 0.77778],
-        "8816": [0.30274, 0.79383, 0, 0, 0.77778],
-        "8817": [0.30274, 0.79383, 0, 0, 0.77778],
-        "8818": [0.22958, 0.72958, 0, 0, 0.77778],
-        "8819": [0.22958, 0.72958, 0, 0, 0.77778],
-        "8822": [0.1808, 0.675, 0, 0, 0.77778],
-        "8823": [0.1808, 0.675, 0, 0, 0.77778],
-        "8828": [0.13667, 0.63667, 0, 0, 0.77778],
-        "8829": [0.13667, 0.63667, 0, 0, 0.77778],
-        "8830": [0.22958, 0.72958, 0, 0, 0.77778],
-        "8831": [0.22958, 0.72958, 0, 0, 0.77778],
-        "8832": [0.20576, 0.70576, 0, 0, 0.77778],
-        "8833": [0.20576, 0.70576, 0, 0, 0.77778],
-        "8840": [0.30274, 0.79383, 0, 0, 0.77778],
-        "8841": [0.30274, 0.79383, 0, 0, 0.77778],
-        "8842": [0.13597, 0.63597, 0, 0, 0.77778],
-        "8843": [0.13597, 0.63597, 0, 0, 0.77778],
-        "8847": [0.03517, 0.54986, 0, 0, 0.77778],
-        "8848": [0.03517, 0.54986, 0, 0, 0.77778],
-        "8858": [0.08198, 0.58198, 0, 0, 0.77778],
-        "8859": [0.08198, 0.58198, 0, 0, 0.77778],
-        "8861": [0.08198, 0.58198, 0, 0, 0.77778],
-        "8862": [0, 0.675, 0, 0, 0.77778],
-        "8863": [0, 0.675, 0, 0, 0.77778],
-        "8864": [0, 0.675, 0, 0, 0.77778],
-        "8865": [0, 0.675, 0, 0, 0.77778],
-        "8872": [0, 0.69224, 0, 0, 0.61111],
-        "8873": [0, 0.69224, 0, 0, 0.72222],
-        "8874": [0, 0.69224, 0, 0, 0.88889],
-        "8876": [0, 0.68889, 0, 0, 0.61111],
-        "8877": [0, 0.68889, 0, 0, 0.61111],
-        "8878": [0, 0.68889, 0, 0, 0.72222],
-        "8879": [0, 0.68889, 0, 0, 0.72222],
-        "8882": [0.03517, 0.54986, 0, 0, 0.77778],
-        "8883": [0.03517, 0.54986, 0, 0, 0.77778],
-        "8884": [0.13667, 0.63667, 0, 0, 0.77778],
-        "8885": [0.13667, 0.63667, 0, 0, 0.77778],
-        "8888": [0, 0.54986, 0, 0, 1.11111],
-        "8890": [0.19444, 0.43056, 0, 0, 0.55556],
-        "8891": [0.19444, 0.69224, 0, 0, 0.61111],
-        "8892": [0.19444, 0.69224, 0, 0, 0.61111],
-        "8901": [0, 0.54986, 0, 0, 0.27778],
-        "8903": [0.08167, 0.58167, 0, 0, 0.77778],
-        "8905": [0.08167, 0.58167, 0, 0, 0.77778],
-        "8906": [0.08167, 0.58167, 0, 0, 0.77778],
-        "8907": [0, 0.69224, 0, 0, 0.77778],
-        "8908": [0, 0.69224, 0, 0, 0.77778],
-        "8909": [-0.03598, 0.46402, 0, 0, 0.77778],
-        "8910": [0, 0.54986, 0, 0, 0.76042],
-        "8911": [0, 0.54986, 0, 0, 0.76042],
-        "8912": [0.03517, 0.54986, 0, 0, 0.77778],
-        "8913": [0.03517, 0.54986, 0, 0, 0.77778],
-        "8914": [0, 0.54986, 0, 0, 0.66667],
-        "8915": [0, 0.54986, 0, 0, 0.66667],
-        "8916": [0, 0.69224, 0, 0, 0.66667],
-        "8918": [0.0391, 0.5391, 0, 0, 0.77778],
-        "8919": [0.0391, 0.5391, 0, 0, 0.77778],
-        "8920": [0.03517, 0.54986, 0, 0, 1.33334],
-        "8921": [0.03517, 0.54986, 0, 0, 1.33334],
-        "8922": [0.38569, 0.88569, 0, 0, 0.77778],
-        "8923": [0.38569, 0.88569, 0, 0, 0.77778],
-        "8926": [0.13667, 0.63667, 0, 0, 0.77778],
-        "8927": [0.13667, 0.63667, 0, 0, 0.77778],
-        "8928": [0.30274, 0.79383, 0, 0, 0.77778],
-        "8929": [0.30274, 0.79383, 0, 0, 0.77778],
-        "8934": [0.23222, 0.74111, 0, 0, 0.77778],
-        "8935": [0.23222, 0.74111, 0, 0, 0.77778],
-        "8936": [0.23222, 0.74111, 0, 0, 0.77778],
-        "8937": [0.23222, 0.74111, 0, 0, 0.77778],
-        "8938": [0.20576, 0.70576, 0, 0, 0.77778],
-        "8939": [0.20576, 0.70576, 0, 0, 0.77778],
-        "8940": [0.30274, 0.79383, 0, 0, 0.77778],
-        "8941": [0.30274, 0.79383, 0, 0, 0.77778],
-        "8994": [0.19444, 0.69224, 0, 0, 0.77778],
-        "8995": [0.19444, 0.69224, 0, 0, 0.77778],
-        "9416": [0.15559, 0.69224, 0, 0, 0.90222],
-        "9484": [0, 0.69224, 0, 0, 0.5],
-        "9488": [0, 0.69224, 0, 0, 0.5],
-        "9492": [0, 0.37788, 0, 0, 0.5],
-        "9496": [0, 0.37788, 0, 0, 0.5],
-        "9585": [0.19444, 0.68889, 0, 0, 0.88889],
-        "9586": [0.19444, 0.74111, 0, 0, 0.88889],
-        "9632": [0, 0.675, 0, 0, 0.77778],
-        "9633": [0, 0.675, 0, 0, 0.77778],
-        "9650": [0, 0.54986, 0, 0, 0.72222],
-        "9651": [0, 0.54986, 0, 0, 0.72222],
-        "9654": [0.03517, 0.54986, 0, 0, 0.77778],
-        "9660": [0, 0.54986, 0, 0, 0.72222],
-        "9661": [0, 0.54986, 0, 0, 0.72222],
-        "9664": [0.03517, 0.54986, 0, 0, 0.77778],
-        "9674": [0.11111, 0.69224, 0, 0, 0.66667],
-        "9733": [0.19444, 0.69224, 0, 0, 0.94445],
-        "10003": [0, 0.69224, 0, 0, 0.83334],
-        "10016": [0, 0.69224, 0, 0, 0.83334],
-        "10731": [0.11111, 0.69224, 0, 0, 0.66667],
-        "10846": [0.19444, 0.75583, 0, 0, 0.61111],
-        "10877": [0.13667, 0.63667, 0, 0, 0.77778],
-        "10878": [0.13667, 0.63667, 0, 0, 0.77778],
-        "10885": [0.25583, 0.75583, 0, 0, 0.77778],
-        "10886": [0.25583, 0.75583, 0, 0, 0.77778],
-        "10887": [0.13597, 0.63597, 0, 0, 0.77778],
-        "10888": [0.13597, 0.63597, 0, 0, 0.77778],
-        "10889": [0.26167, 0.75726, 0, 0, 0.77778],
-        "10890": [0.26167, 0.75726, 0, 0, 0.77778],
-        "10891": [0.48256, 0.98256, 0, 0, 0.77778],
-        "10892": [0.48256, 0.98256, 0, 0, 0.77778],
-        "10901": [0.13667, 0.63667, 0, 0, 0.77778],
-        "10902": [0.13667, 0.63667, 0, 0, 0.77778],
-        "10933": [0.25142, 0.75726, 0, 0, 0.77778],
-        "10934": [0.25142, 0.75726, 0, 0, 0.77778],
-        "10935": [0.26167, 0.75726, 0, 0, 0.77778],
-        "10936": [0.26167, 0.75726, 0, 0, 0.77778],
-        "10937": [0.26167, 0.75726, 0, 0, 0.77778],
-        "10938": [0.26167, 0.75726, 0, 0, 0.77778],
-        "10949": [0.25583, 0.75583, 0, 0, 0.77778],
-        "10950": [0.25583, 0.75583, 0, 0, 0.77778],
-        "10955": [0.28481, 0.79383, 0, 0, 0.77778],
-        "10956": [0.28481, 0.79383, 0, 0, 0.77778],
-        "57350": [0.08167, 0.58167, 0, 0, 0.22222],
-        "57351": [0.08167, 0.58167, 0, 0, 0.38889],
-        "57352": [0.08167, 0.58167, 0, 0, 0.77778],
-        "57353": [0, 0.43056, 0.04028, 0, 0.66667],
-        "57356": [0.25142, 0.75726, 0, 0, 0.77778],
-        "57357": [0.25142, 0.75726, 0, 0, 0.77778],
-        "57358": [0.41951, 0.91951, 0, 0, 0.77778],
-        "57359": [0.30274, 0.79383, 0, 0, 0.77778],
-        "57360": [0.30274, 0.79383, 0, 0, 0.77778],
-        "57361": [0.41951, 0.91951, 0, 0, 0.77778],
-        "57366": [0.25142, 0.75726, 0, 0, 0.77778],
-        "57367": [0.25142, 0.75726, 0, 0, 0.77778],
-        "57368": [0.25142, 0.75726, 0, 0, 0.77778],
-        "57369": [0.25142, 0.75726, 0, 0, 0.77778],
-        "57370": [0.13597, 0.63597, 0, 0, 0.77778],
-        "57371": [0.13597, 0.63597, 0, 0, 0.77778],
-    },
-    "Caligraphic-Regular": {
-        "48": [0, 0.43056, 0, 0, 0.5],
-        "49": [0, 0.43056, 0, 0, 0.5],
-        "50": [0, 0.43056, 0, 0, 0.5],
-        "51": [0.19444, 0.43056, 0, 0, 0.5],
-        "52": [0.19444, 0.43056, 0, 0, 0.5],
-        "53": [0.19444, 0.43056, 0, 0, 0.5],
-        "54": [0, 0.64444, 0, 0, 0.5],
-        "55": [0.19444, 0.43056, 0, 0, 0.5],
-        "56": [0, 0.64444, 0, 0, 0.5],
-        "57": [0.19444, 0.43056, 0, 0, 0.5],
-        "65": [0, 0.68333, 0, 0.19445, 0.79847],
-        "66": [0, 0.68333, 0.03041, 0.13889, 0.65681],
-        "67": [0, 0.68333, 0.05834, 0.13889, 0.52653],
-        "68": [0, 0.68333, 0.02778, 0.08334, 0.77139],
-        "69": [0, 0.68333, 0.08944, 0.11111, 0.52778],
-        "70": [0, 0.68333, 0.09931, 0.11111, 0.71875],
-        "71": [0.09722, 0.68333, 0.0593, 0.11111, 0.59487],
-        "72": [0, 0.68333, 0.00965, 0.11111, 0.84452],
-        "73": [0, 0.68333, 0.07382, 0, 0.54452],
-        "74": [0.09722, 0.68333, 0.18472, 0.16667, 0.67778],
-        "75": [0, 0.68333, 0.01445, 0.05556, 0.76195],
-        "76": [0, 0.68333, 0, 0.13889, 0.68972],
-        "77": [0, 0.68333, 0, 0.13889, 1.2009],
-        "78": [0, 0.68333, 0.14736, 0.08334, 0.82049],
-        "79": [0, 0.68333, 0.02778, 0.11111, 0.79611],
-        "80": [0, 0.68333, 0.08222, 0.08334, 0.69556],
-        "81": [0.09722, 0.68333, 0, 0.11111, 0.81667],
-        "82": [0, 0.68333, 0, 0.08334, 0.8475],
-        "83": [0, 0.68333, 0.075, 0.13889, 0.60556],
-        "84": [0, 0.68333, 0.25417, 0, 0.54464],
-        "85": [0, 0.68333, 0.09931, 0.08334, 0.62583],
-        "86": [0, 0.68333, 0.08222, 0, 0.61278],
-        "87": [0, 0.68333, 0.08222, 0.08334, 0.98778],
-        "88": [0, 0.68333, 0.14643, 0.13889, 0.7133],
-        "89": [0.09722, 0.68333, 0.08222, 0.08334, 0.66834],
-        "90": [0, 0.68333, 0.07944, 0.13889, 0.72473],
-    },
-    "Fraktur-Regular": {
-        "33": [0, 0.69141, 0, 0, 0.29574],
-        "34": [0, 0.69141, 0, 0, 0.21471],
-        "38": [0, 0.69141, 0, 0, 0.73786],
-        "39": [0, 0.69141, 0, 0, 0.21201],
-        "40": [0.24982, 0.74947, 0, 0, 0.38865],
-        "41": [0.24982, 0.74947, 0, 0, 0.38865],
-        "42": [0, 0.62119, 0, 0, 0.27764],
-        "43": [0.08319, 0.58283, 0, 0, 0.75623],
-        "44": [0, 0.10803, 0, 0, 0.27764],
-        "45": [0.08319, 0.58283, 0, 0, 0.75623],
-        "46": [0, 0.10803, 0, 0, 0.27764],
-        "47": [0.24982, 0.74947, 0, 0, 0.50181],
-        "48": [0, 0.47534, 0, 0, 0.50181],
-        "49": [0, 0.47534, 0, 0, 0.50181],
-        "50": [0, 0.47534, 0, 0, 0.50181],
-        "51": [0.18906, 0.47534, 0, 0, 0.50181],
-        "52": [0.18906, 0.47534, 0, 0, 0.50181],
-        "53": [0.18906, 0.47534, 0, 0, 0.50181],
-        "54": [0, 0.69141, 0, 0, 0.50181],
-        "55": [0.18906, 0.47534, 0, 0, 0.50181],
-        "56": [0, 0.69141, 0, 0, 0.50181],
-        "57": [0.18906, 0.47534, 0, 0, 0.50181],
-        "58": [0, 0.47534, 0, 0, 0.21606],
-        "59": [0.12604, 0.47534, 0, 0, 0.21606],
-        "61": [-0.13099, 0.36866, 0, 0, 0.75623],
-        "63": [0, 0.69141, 0, 0, 0.36245],
-        "65": [0, 0.69141, 0, 0, 0.7176],
-        "66": [0, 0.69141, 0, 0, 0.88397],
-        "67": [0, 0.69141, 0, 0, 0.61254],
-        "68": [0, 0.69141, 0, 0, 0.83158],
-        "69": [0, 0.69141, 0, 0, 0.66278],
-        "70": [0.12604, 0.69141, 0, 0, 0.61119],
-        "71": [0, 0.69141, 0, 0, 0.78539],
-        "72": [0.06302, 0.69141, 0, 0, 0.7203],
-        "73": [0, 0.69141, 0, 0, 0.55448],
-        "74": [0.12604, 0.69141, 0, 0, 0.55231],
-        "75": [0, 0.69141, 0, 0, 0.66845],
-        "76": [0, 0.69141, 0, 0, 0.66602],
-        "77": [0, 0.69141, 0, 0, 1.04953],
-        "78": [0, 0.69141, 0, 0, 0.83212],
-        "79": [0, 0.69141, 0, 0, 0.82699],
-        "80": [0.18906, 0.69141, 0, 0, 0.82753],
-        "81": [0.03781, 0.69141, 0, 0, 0.82699],
-        "82": [0, 0.69141, 0, 0, 0.82807],
-        "83": [0, 0.69141, 0, 0, 0.82861],
-        "84": [0, 0.69141, 0, 0, 0.66899],
-        "85": [0, 0.69141, 0, 0, 0.64576],
-        "86": [0, 0.69141, 0, 0, 0.83131],
-        "87": [0, 0.69141, 0, 0, 1.04602],
-        "88": [0, 0.69141, 0, 0, 0.71922],
-        "89": [0.18906, 0.69141, 0, 0, 0.83293],
-        "90": [0.12604, 0.69141, 0, 0, 0.60201],
-        "91": [0.24982, 0.74947, 0, 0, 0.27764],
-        "93": [0.24982, 0.74947, 0, 0, 0.27764],
-        "94": [0, 0.69141, 0, 0, 0.49965],
-        "97": [0, 0.47534, 0, 0, 0.50046],
-        "98": [0, 0.69141, 0, 0, 0.51315],
-        "99": [0, 0.47534, 0, 0, 0.38946],
-        "100": [0, 0.62119, 0, 0, 0.49857],
-        "101": [0, 0.47534, 0, 0, 0.40053],
-        "102": [0.18906, 0.69141, 0, 0, 0.32626],
-        "103": [0.18906, 0.47534, 0, 0, 0.5037],
-        "104": [0.18906, 0.69141, 0, 0, 0.52126],
-        "105": [0, 0.69141, 0, 0, 0.27899],
-        "106": [0, 0.69141, 0, 0, 0.28088],
-        "107": [0, 0.69141, 0, 0, 0.38946],
-        "108": [0, 0.69141, 0, 0, 0.27953],
-        "109": [0, 0.47534, 0, 0, 0.76676],
-        "110": [0, 0.47534, 0, 0, 0.52666],
-        "111": [0, 0.47534, 0, 0, 0.48885],
-        "112": [0.18906, 0.52396, 0, 0, 0.50046],
-        "113": [0.18906, 0.47534, 0, 0, 0.48912],
-        "114": [0, 0.47534, 0, 0, 0.38919],
-        "115": [0, 0.47534, 0, 0, 0.44266],
-        "116": [0, 0.62119, 0, 0, 0.33301],
-        "117": [0, 0.47534, 0, 0, 0.5172],
-        "118": [0, 0.52396, 0, 0, 0.5118],
-        "119": [0, 0.52396, 0, 0, 0.77351],
-        "120": [0.18906, 0.47534, 0, 0, 0.38865],
-        "121": [0.18906, 0.47534, 0, 0, 0.49884],
-        "122": [0.18906, 0.47534, 0, 0, 0.39054],
-        "8216": [0, 0.69141, 0, 0, 0.21471],
-        "8217": [0, 0.69141, 0, 0, 0.21471],
-        "58112": [0, 0.62119, 0, 0, 0.49749],
-        "58113": [0, 0.62119, 0, 0, 0.4983],
-        "58114": [0.18906, 0.69141, 0, 0, 0.33328],
-        "58115": [0.18906, 0.69141, 0, 0, 0.32923],
-        "58116": [0.18906, 0.47534, 0, 0, 0.50343],
-        "58117": [0, 0.69141, 0, 0, 0.33301],
-        "58118": [0, 0.62119, 0, 0, 0.33409],
-        "58119": [0, 0.47534, 0, 0, 0.50073],
-    },
-    "Main-Bold": {
-        "33": [0, 0.69444, 0, 0, 0.35],
-        "34": [0, 0.69444, 0, 0, 0.60278],
-        "35": [0.19444, 0.69444, 0, 0, 0.95833],
-        "36": [0.05556, 0.75, 0, 0, 0.575],
-        "37": [0.05556, 0.75, 0, 0, 0.95833],
-        "38": [0, 0.69444, 0, 0, 0.89444],
-        "39": [0, 0.69444, 0, 0, 0.31944],
-        "40": [0.25, 0.75, 0, 0, 0.44722],
-        "41": [0.25, 0.75, 0, 0, 0.44722],
-        "42": [0, 0.75, 0, 0, 0.575],
-        "43": [0.13333, 0.63333, 0, 0, 0.89444],
-        "44": [0.19444, 0.15556, 0, 0, 0.31944],
-        "45": [0, 0.44444, 0, 0, 0.38333],
-        "46": [0, 0.15556, 0, 0, 0.31944],
-        "47": [0.25, 0.75, 0, 0, 0.575],
-        "48": [0, 0.64444, 0, 0, 0.575],
-        "49": [0, 0.64444, 0, 0, 0.575],
-        "50": [0, 0.64444, 0, 0, 0.575],
-        "51": [0, 0.64444, 0, 0, 0.575],
-        "52": [0, 0.64444, 0, 0, 0.575],
-        "53": [0, 0.64444, 0, 0, 0.575],
-        "54": [0, 0.64444, 0, 0, 0.575],
-        "55": [0, 0.64444, 0, 0, 0.575],
-        "56": [0, 0.64444, 0, 0, 0.575],
-        "57": [0, 0.64444, 0, 0, 0.575],
-        "58": [0, 0.44444, 0, 0, 0.31944],
-        "59": [0.19444, 0.44444, 0, 0, 0.31944],
-        "60": [0.08556, 0.58556, 0, 0, 0.89444],
-        "61": [-0.10889, 0.39111, 0, 0, 0.89444],
-        "62": [0.08556, 0.58556, 0, 0, 0.89444],
-        "63": [0, 0.69444, 0, 0, 0.54305],
-        "64": [0, 0.69444, 0, 0, 0.89444],
-        "65": [0, 0.68611, 0, 0, 0.86944],
-        "66": [0, 0.68611, 0, 0, 0.81805],
-        "67": [0, 0.68611, 0, 0, 0.83055],
-        "68": [0, 0.68611, 0, 0, 0.88194],
-        "69": [0, 0.68611, 0, 0, 0.75555],
-        "70": [0, 0.68611, 0, 0, 0.72361],
-        "71": [0, 0.68611, 0, 0, 0.90416],
-        "72": [0, 0.68611, 0, 0, 0.9],
-        "73": [0, 0.68611, 0, 0, 0.43611],
-        "74": [0, 0.68611, 0, 0, 0.59444],
-        "75": [0, 0.68611, 0, 0, 0.90138],
-        "76": [0, 0.68611, 0, 0, 0.69166],
-        "77": [0, 0.68611, 0, 0, 1.09166],
-        "78": [0, 0.68611, 0, 0, 0.9],
-        "79": [0, 0.68611, 0, 0, 0.86388],
-        "80": [0, 0.68611, 0, 0, 0.78611],
-        "81": [0.19444, 0.68611, 0, 0, 0.86388],
-        "82": [0, 0.68611, 0, 0, 0.8625],
-        "83": [0, 0.68611, 0, 0, 0.63889],
-        "84": [0, 0.68611, 0, 0, 0.8],
-        "85": [0, 0.68611, 0, 0, 0.88472],
-        "86": [0, 0.68611, 0.01597, 0, 0.86944],
-        "87": [0, 0.68611, 0.01597, 0, 1.18888],
-        "88": [0, 0.68611, 0, 0, 0.86944],
-        "89": [0, 0.68611, 0.02875, 0, 0.86944],
-        "90": [0, 0.68611, 0, 0, 0.70277],
-        "91": [0.25, 0.75, 0, 0, 0.31944],
-        "92": [0.25, 0.75, 0, 0, 0.575],
-        "93": [0.25, 0.75, 0, 0, 0.31944],
-        "94": [0, 0.69444, 0, 0, 0.575],
-        "95": [0.31, 0.13444, 0.03194, 0, 0.575],
-        "97": [0, 0.44444, 0, 0, 0.55902],
-        "98": [0, 0.69444, 0, 0, 0.63889],
-        "99": [0, 0.44444, 0, 0, 0.51111],
-        "100": [0, 0.69444, 0, 0, 0.63889],
-        "101": [0, 0.44444, 0, 0, 0.52708],
-        "102": [0, 0.69444, 0.10903, 0, 0.35139],
-        "103": [0.19444, 0.44444, 0.01597, 0, 0.575],
-        "104": [0, 0.69444, 0, 0, 0.63889],
-        "105": [0, 0.69444, 0, 0, 0.31944],
-        "106": [0.19444, 0.69444, 0, 0, 0.35139],
-        "107": [0, 0.69444, 0, 0, 0.60694],
-        "108": [0, 0.69444, 0, 0, 0.31944],
-        "109": [0, 0.44444, 0, 0, 0.95833],
-        "110": [0, 0.44444, 0, 0, 0.63889],
-        "111": [0, 0.44444, 0, 0, 0.575],
-        "112": [0.19444, 0.44444, 0, 0, 0.63889],
-        "113": [0.19444, 0.44444, 0, 0, 0.60694],
-        "114": [0, 0.44444, 0, 0, 0.47361],
-        "115": [0, 0.44444, 0, 0, 0.45361],
-        "116": [0, 0.63492, 0, 0, 0.44722],
-        "117": [0, 0.44444, 0, 0, 0.63889],
-        "118": [0, 0.44444, 0.01597, 0, 0.60694],
-        "119": [0, 0.44444, 0.01597, 0, 0.83055],
-        "120": [0, 0.44444, 0, 0, 0.60694],
-        "121": [0.19444, 0.44444, 0.01597, 0, 0.60694],
-        "122": [0, 0.44444, 0, 0, 0.51111],
-        "123": [0.25, 0.75, 0, 0, 0.575],
-        "124": [0.25, 0.75, 0, 0, 0.31944],
-        "125": [0.25, 0.75, 0, 0, 0.575],
-        "126": [0.35, 0.34444, 0, 0, 0.575],
-        "168": [0, 0.69444, 0, 0, 0.575],
-        "172": [0, 0.44444, 0, 0, 0.76666],
-        "176": [0, 0.69444, 0, 0, 0.86944],
-        "177": [0.13333, 0.63333, 0, 0, 0.89444],
-        "198": [0, 0.68611, 0, 0, 1.04166],
-        "215": [0.13333, 0.63333, 0, 0, 0.89444],
-        "216": [0.04861, 0.73472, 0, 0, 0.89444],
-        "223": [0, 0.69444, 0, 0, 0.59722],
-        "230": [0, 0.44444, 0, 0, 0.83055],
-        "247": [0.13333, 0.63333, 0, 0, 0.89444],
-        "248": [0.09722, 0.54167, 0, 0, 0.575],
-        "305": [0, 0.44444, 0, 0, 0.31944],
-        "338": [0, 0.68611, 0, 0, 1.16944],
-        "339": [0, 0.44444, 0, 0, 0.89444],
-        "567": [0.19444, 0.44444, 0, 0, 0.35139],
-        "710": [0, 0.69444, 0, 0, 0.575],
-        "711": [0, 0.63194, 0, 0, 0.575],
-        "713": [0, 0.59611, 0, 0, 0.575],
-        "714": [0, 0.69444, 0, 0, 0.575],
-        "715": [0, 0.69444, 0, 0, 0.575],
-        "728": [0, 0.69444, 0, 0, 0.575],
-        "729": [0, 0.69444, 0, 0, 0.31944],
-        "730": [0, 0.69444, 0, 0, 0.86944],
-        "732": [0, 0.69444, 0, 0, 0.575],
-        "733": [0, 0.69444, 0, 0, 0.575],
-        "824": [0.19444, 0.69444, 0, 0, 0],
-        "915": [0, 0.68611, 0, 0, 0.69166],
-        "916": [0, 0.68611, 0, 0, 0.95833],
-        "920": [0, 0.68611, 0, 0, 0.89444],
-        "923": [0, 0.68611, 0, 0, 0.80555],
-        "926": [0, 0.68611, 0, 0, 0.76666],
-        "928": [0, 0.68611, 0, 0, 0.9],
-        "931": [0, 0.68611, 0, 0, 0.83055],
-        "933": [0, 0.68611, 0, 0, 0.89444],
-        "934": [0, 0.68611, 0, 0, 0.83055],
-        "936": [0, 0.68611, 0, 0, 0.89444],
-        "937": [0, 0.68611, 0, 0, 0.83055],
-        "8211": [0, 0.44444, 0.03194, 0, 0.575],
-        "8212": [0, 0.44444, 0.03194, 0, 1.14999],
-        "8216": [0, 0.69444, 0, 0, 0.31944],
-        "8217": [0, 0.69444, 0, 0, 0.31944],
-        "8220": [0, 0.69444, 0, 0, 0.60278],
-        "8221": [0, 0.69444, 0, 0, 0.60278],
-        "8224": [0.19444, 0.69444, 0, 0, 0.51111],
-        "8225": [0.19444, 0.69444, 0, 0, 0.51111],
-        "8242": [0, 0.55556, 0, 0, 0.34444],
-        "8407": [0, 0.72444, 0.15486, 0, 0.575],
-        "8463": [0, 0.69444, 0, 0, 0.66759],
-        "8465": [0, 0.69444, 0, 0, 0.83055],
-        "8467": [0, 0.69444, 0, 0, 0.47361],
-        "8472": [0.19444, 0.44444, 0, 0, 0.74027],
-        "8476": [0, 0.69444, 0, 0, 0.83055],
-        "8501": [0, 0.69444, 0, 0, 0.70277],
-        "8592": [-0.10889, 0.39111, 0, 0, 1.14999],
-        "8593": [0.19444, 0.69444, 0, 0, 0.575],
-        "8594": [-0.10889, 0.39111, 0, 0, 1.14999],
-        "8595": [0.19444, 0.69444, 0, 0, 0.575],
-        "8596": [-0.10889, 0.39111, 0, 0, 1.14999],
-        "8597": [0.25, 0.75, 0, 0, 0.575],
-        "8598": [0.19444, 0.69444, 0, 0, 1.14999],
-        "8599": [0.19444, 0.69444, 0, 0, 1.14999],
-        "8600": [0.19444, 0.69444, 0, 0, 1.14999],
-        "8601": [0.19444, 0.69444, 0, 0, 1.14999],
-        "8636": [-0.10889, 0.39111, 0, 0, 1.14999],
-        "8637": [-0.10889, 0.39111, 0, 0, 1.14999],
-        "8640": [-0.10889, 0.39111, 0, 0, 1.14999],
-        "8641": [-0.10889, 0.39111, 0, 0, 1.14999],
-        "8656": [-0.10889, 0.39111, 0, 0, 1.14999],
-        "8657": [0.19444, 0.69444, 0, 0, 0.70277],
-        "8658": [-0.10889, 0.39111, 0, 0, 1.14999],
-        "8659": [0.19444, 0.69444, 0, 0, 0.70277],
-        "8660": [-0.10889, 0.39111, 0, 0, 1.14999],
-        "8661": [0.25, 0.75, 0, 0, 0.70277],
-        "8704": [0, 0.69444, 0, 0, 0.63889],
-        "8706": [0, 0.69444, 0.06389, 0, 0.62847],
-        "8707": [0, 0.69444, 0, 0, 0.63889],
-        "8709": [0.05556, 0.75, 0, 0, 0.575],
-        "8711": [0, 0.68611, 0, 0, 0.95833],
-        "8712": [0.08556, 0.58556, 0, 0, 0.76666],
-        "8715": [0.08556, 0.58556, 0, 0, 0.76666],
-        "8722": [0.13333, 0.63333, 0, 0, 0.89444],
-        "8723": [0.13333, 0.63333, 0, 0, 0.89444],
-        "8725": [0.25, 0.75, 0, 0, 0.575],
-        "8726": [0.25, 0.75, 0, 0, 0.575],
-        "8727": [-0.02778, 0.47222, 0, 0, 0.575],
-        "8728": [-0.02639, 0.47361, 0, 0, 0.575],
-        "8729": [-0.02639, 0.47361, 0, 0, 0.575],
-        "8730": [0.18, 0.82, 0, 0, 0.95833],
-        "8733": [0, 0.44444, 0, 0, 0.89444],
-        "8734": [0, 0.44444, 0, 0, 1.14999],
-        "8736": [0, 0.69224, 0, 0, 0.72222],
-        "8739": [0.25, 0.75, 0, 0, 0.31944],
-        "8741": [0.25, 0.75, 0, 0, 0.575],
-        "8743": [0, 0.55556, 0, 0, 0.76666],
-        "8744": [0, 0.55556, 0, 0, 0.76666],
-        "8745": [0, 0.55556, 0, 0, 0.76666],
-        "8746": [0, 0.55556, 0, 0, 0.76666],
-        "8747": [0.19444, 0.69444, 0.12778, 0, 0.56875],
-        "8764": [-0.10889, 0.39111, 0, 0, 0.89444],
-        "8768": [0.19444, 0.69444, 0, 0, 0.31944],
-        "8771": [0.00222, 0.50222, 0, 0, 0.89444],
-        "8776": [0.02444, 0.52444, 0, 0, 0.89444],
-        "8781": [0.00222, 0.50222, 0, 0, 0.89444],
-        "8801": [0.00222, 0.50222, 0, 0, 0.89444],
-        "8804": [0.19667, 0.69667, 0, 0, 0.89444],
-        "8805": [0.19667, 0.69667, 0, 0, 0.89444],
-        "8810": [0.08556, 0.58556, 0, 0, 1.14999],
-        "8811": [0.08556, 0.58556, 0, 0, 1.14999],
-        "8826": [0.08556, 0.58556, 0, 0, 0.89444],
-        "8827": [0.08556, 0.58556, 0, 0, 0.89444],
-        "8834": [0.08556, 0.58556, 0, 0, 0.89444],
-        "8835": [0.08556, 0.58556, 0, 0, 0.89444],
-        "8838": [0.19667, 0.69667, 0, 0, 0.89444],
-        "8839": [0.19667, 0.69667, 0, 0, 0.89444],
-        "8846": [0, 0.55556, 0, 0, 0.76666],
-        "8849": [0.19667, 0.69667, 0, 0, 0.89444],
-        "8850": [0.19667, 0.69667, 0, 0, 0.89444],
-        "8851": [0, 0.55556, 0, 0, 0.76666],
-        "8852": [0, 0.55556, 0, 0, 0.76666],
-        "8853": [0.13333, 0.63333, 0, 0, 0.89444],
-        "8854": [0.13333, 0.63333, 0, 0, 0.89444],
-        "8855": [0.13333, 0.63333, 0, 0, 0.89444],
-        "8856": [0.13333, 0.63333, 0, 0, 0.89444],
-        "8857": [0.13333, 0.63333, 0, 0, 0.89444],
-        "8866": [0, 0.69444, 0, 0, 0.70277],
-        "8867": [0, 0.69444, 0, 0, 0.70277],
-        "8868": [0, 0.69444, 0, 0, 0.89444],
-        "8869": [0, 0.69444, 0, 0, 0.89444],
-        "8900": [-0.02639, 0.47361, 0, 0, 0.575],
-        "8901": [-0.02639, 0.47361, 0, 0, 0.31944],
-        "8902": [-0.02778, 0.47222, 0, 0, 0.575],
-        "8968": [0.25, 0.75, 0, 0, 0.51111],
-        "8969": [0.25, 0.75, 0, 0, 0.51111],
-        "8970": [0.25, 0.75, 0, 0, 0.51111],
-        "8971": [0.25, 0.75, 0, 0, 0.51111],
-        "8994": [-0.13889, 0.36111, 0, 0, 1.14999],
-        "8995": [-0.13889, 0.36111, 0, 0, 1.14999],
-        "9651": [0.19444, 0.69444, 0, 0, 1.02222],
-        "9657": [-0.02778, 0.47222, 0, 0, 0.575],
-        "9661": [0.19444, 0.69444, 0, 0, 1.02222],
-        "9667": [-0.02778, 0.47222, 0, 0, 0.575],
-        "9711": [0.19444, 0.69444, 0, 0, 1.14999],
-        "9824": [0.12963, 0.69444, 0, 0, 0.89444],
-        "9825": [0.12963, 0.69444, 0, 0, 0.89444],
-        "9826": [0.12963, 0.69444, 0, 0, 0.89444],
-        "9827": [0.12963, 0.69444, 0, 0, 0.89444],
-        "9837": [0, 0.75, 0, 0, 0.44722],
-        "9838": [0.19444, 0.69444, 0, 0, 0.44722],
-        "9839": [0.19444, 0.69444, 0, 0, 0.44722],
-        "10216": [0.25, 0.75, 0, 0, 0.44722],
-        "10217": [0.25, 0.75, 0, 0, 0.44722],
-        "10815": [0, 0.68611, 0, 0, 0.9],
-        "10927": [0.19667, 0.69667, 0, 0, 0.89444],
-        "10928": [0.19667, 0.69667, 0, 0, 0.89444],
-    },
-    "Main-BoldItalic": {
-        "33": [0, 0.69444, 0.11417, 0, 0.38611],
-        "34": [0, 0.69444, 0.07939, 0, 0.62055],
-        "35": [0.19444, 0.69444, 0.06833, 0, 0.94444],
-        "37": [0.05556, 0.75, 0.12861, 0, 0.94444],
-        "38": [0, 0.69444, 0.08528, 0, 0.88555],
-        "39": [0, 0.69444, 0.12945, 0, 0.35555],
-        "40": [0.25, 0.75, 0.15806, 0, 0.47333],
-        "41": [0.25, 0.75, 0.03306, 0, 0.47333],
-        "42": [0, 0.75, 0.14333, 0, 0.59111],
-        "43": [0.10333, 0.60333, 0.03306, 0, 0.88555],
-        "44": [0.19444, 0.14722, 0, 0, 0.35555],
-        "45": [0, 0.44444, 0.02611, 0, 0.41444],
-        "46": [0, 0.14722, 0, 0, 0.35555],
-        "47": [0.25, 0.75, 0.15806, 0, 0.59111],
-        "48": [0, 0.64444, 0.13167, 0, 0.59111],
-        "49": [0, 0.64444, 0.13167, 0, 0.59111],
-        "50": [0, 0.64444, 0.13167, 0, 0.59111],
-        "51": [0, 0.64444, 0.13167, 0, 0.59111],
-        "52": [0.19444, 0.64444, 0.13167, 0, 0.59111],
-        "53": [0, 0.64444, 0.13167, 0, 0.59111],
-        "54": [0, 0.64444, 0.13167, 0, 0.59111],
-        "55": [0.19444, 0.64444, 0.13167, 0, 0.59111],
-        "56": [0, 0.64444, 0.13167, 0, 0.59111],
-        "57": [0, 0.64444, 0.13167, 0, 0.59111],
-        "58": [0, 0.44444, 0.06695, 0, 0.35555],
-        "59": [0.19444, 0.44444, 0.06695, 0, 0.35555],
-        "61": [-0.10889, 0.39111, 0.06833, 0, 0.88555],
-        "63": [0, 0.69444, 0.11472, 0, 0.59111],
-        "64": [0, 0.69444, 0.09208, 0, 0.88555],
-        "65": [0, 0.68611, 0, 0, 0.86555],
-        "66": [0, 0.68611, 0.0992, 0, 0.81666],
-        "67": [0, 0.68611, 0.14208, 0, 0.82666],
-        "68": [0, 0.68611, 0.09062, 0, 0.87555],
-        "69": [0, 0.68611, 0.11431, 0, 0.75666],
-        "70": [0, 0.68611, 0.12903, 0, 0.72722],
-        "71": [0, 0.68611, 0.07347, 0, 0.89527],
-        "72": [0, 0.68611, 0.17208, 0, 0.8961],
-        "73": [0, 0.68611, 0.15681, 0, 0.47166],
-        "74": [0, 0.68611, 0.145, 0, 0.61055],
-        "75": [0, 0.68611, 0.14208, 0, 0.89499],
-        "76": [0, 0.68611, 0, 0, 0.69777],
-        "77": [0, 0.68611, 0.17208, 0, 1.07277],
-        "78": [0, 0.68611, 0.17208, 0, 0.8961],
-        "79": [0, 0.68611, 0.09062, 0, 0.85499],
-        "80": [0, 0.68611, 0.0992, 0, 0.78721],
-        "81": [0.19444, 0.68611, 0.09062, 0, 0.85499],
-        "82": [0, 0.68611, 0.02559, 0, 0.85944],
-        "83": [0, 0.68611, 0.11264, 0, 0.64999],
-        "84": [0, 0.68611, 0.12903, 0, 0.7961],
-        "85": [0, 0.68611, 0.17208, 0, 0.88083],
-        "86": [0, 0.68611, 0.18625, 0, 0.86555],
-        "87": [0, 0.68611, 0.18625, 0, 1.15999],
-        "88": [0, 0.68611, 0.15681, 0, 0.86555],
-        "89": [0, 0.68611, 0.19803, 0, 0.86555],
-        "90": [0, 0.68611, 0.14208, 0, 0.70888],
-        "91": [0.25, 0.75, 0.1875, 0, 0.35611],
-        "93": [0.25, 0.75, 0.09972, 0, 0.35611],
-        "94": [0, 0.69444, 0.06709, 0, 0.59111],
-        "95": [0.31, 0.13444, 0.09811, 0, 0.59111],
-        "97": [0, 0.44444, 0.09426, 0, 0.59111],
-        "98": [0, 0.69444, 0.07861, 0, 0.53222],
-        "99": [0, 0.44444, 0.05222, 0, 0.53222],
-        "100": [0, 0.69444, 0.10861, 0, 0.59111],
-        "101": [0, 0.44444, 0.085, 0, 0.53222],
-        "102": [0.19444, 0.69444, 0.21778, 0, 0.4],
-        "103": [0.19444, 0.44444, 0.105, 0, 0.53222],
-        "104": [0, 0.69444, 0.09426, 0, 0.59111],
-        "105": [0, 0.69326, 0.11387, 0, 0.35555],
-        "106": [0.19444, 0.69326, 0.1672, 0, 0.35555],
-        "107": [0, 0.69444, 0.11111, 0, 0.53222],
-        "108": [0, 0.69444, 0.10861, 0, 0.29666],
-        "109": [0, 0.44444, 0.09426, 0, 0.94444],
-        "110": [0, 0.44444, 0.09426, 0, 0.64999],
-        "111": [0, 0.44444, 0.07861, 0, 0.59111],
-        "112": [0.19444, 0.44444, 0.07861, 0, 0.59111],
-        "113": [0.19444, 0.44444, 0.105, 0, 0.53222],
-        "114": [0, 0.44444, 0.11111, 0, 0.50167],
-        "115": [0, 0.44444, 0.08167, 0, 0.48694],
-        "116": [0, 0.63492, 0.09639, 0, 0.385],
-        "117": [0, 0.44444, 0.09426, 0, 0.62055],
-        "118": [0, 0.44444, 0.11111, 0, 0.53222],
-        "119": [0, 0.44444, 0.11111, 0, 0.76777],
-        "120": [0, 0.44444, 0.12583, 0, 0.56055],
-        "121": [0.19444, 0.44444, 0.105, 0, 0.56166],
-        "122": [0, 0.44444, 0.13889, 0, 0.49055],
-        "126": [0.35, 0.34444, 0.11472, 0, 0.59111],
-        "163": [0, 0.69444, 0, 0, 0.86853],
-        "168": [0, 0.69444, 0.11473, 0, 0.59111],
-        "176": [0, 0.69444, 0, 0, 0.94888],
-        "198": [0, 0.68611, 0.11431, 0, 1.02277],
-        "216": [0.04861, 0.73472, 0.09062, 0, 0.88555],
-        "223": [0.19444, 0.69444, 0.09736, 0, 0.665],
-        "230": [0, 0.44444, 0.085, 0, 0.82666],
-        "248": [0.09722, 0.54167, 0.09458, 0, 0.59111],
-        "305": [0, 0.44444, 0.09426, 0, 0.35555],
-        "338": [0, 0.68611, 0.11431, 0, 1.14054],
-        "339": [0, 0.44444, 0.085, 0, 0.82666],
-        "567": [0.19444, 0.44444, 0.04611, 0, 0.385],
-        "710": [0, 0.69444, 0.06709, 0, 0.59111],
-        "711": [0, 0.63194, 0.08271, 0, 0.59111],
-        "713": [0, 0.59444, 0.10444, 0, 0.59111],
-        "714": [0, 0.69444, 0.08528, 0, 0.59111],
-        "715": [0, 0.69444, 0, 0, 0.59111],
-        "728": [0, 0.69444, 0.10333, 0, 0.59111],
-        "729": [0, 0.69444, 0.12945, 0, 0.35555],
-        "730": [0, 0.69444, 0, 0, 0.94888],
-        "732": [0, 0.69444, 0.11472, 0, 0.59111],
-        "733": [0, 0.69444, 0.11472, 0, 0.59111],
-        "915": [0, 0.68611, 0.12903, 0, 0.69777],
-        "916": [0, 0.68611, 0, 0, 0.94444],
-        "920": [0, 0.68611, 0.09062, 0, 0.88555],
-        "923": [0, 0.68611, 0, 0, 0.80666],
-        "926": [0, 0.68611, 0.15092, 0, 0.76777],
-        "928": [0, 0.68611, 0.17208, 0, 0.8961],
-        "931": [0, 0.68611, 0.11431, 0, 0.82666],
-        "933": [0, 0.68611, 0.10778, 0, 0.88555],
-        "934": [0, 0.68611, 0.05632, 0, 0.82666],
-        "936": [0, 0.68611, 0.10778, 0, 0.88555],
-        "937": [0, 0.68611, 0.0992, 0, 0.82666],
-        "8211": [0, 0.44444, 0.09811, 0, 0.59111],
-        "8212": [0, 0.44444, 0.09811, 0, 1.18221],
-        "8216": [0, 0.69444, 0.12945, 0, 0.35555],
-        "8217": [0, 0.69444, 0.12945, 0, 0.35555],
-        "8220": [0, 0.69444, 0.16772, 0, 0.62055],
-        "8221": [0, 0.69444, 0.07939, 0, 0.62055],
-    },
-    "Main-Italic": {
-        "33": [0, 0.69444, 0.12417, 0, 0.30667],
-        "34": [0, 0.69444, 0.06961, 0, 0.51444],
-        "35": [0.19444, 0.69444, 0.06616, 0, 0.81777],
-        "37": [0.05556, 0.75, 0.13639, 0, 0.81777],
-        "38": [0, 0.69444, 0.09694, 0, 0.76666],
-        "39": [0, 0.69444, 0.12417, 0, 0.30667],
-        "40": [0.25, 0.75, 0.16194, 0, 0.40889],
-        "41": [0.25, 0.75, 0.03694, 0, 0.40889],
-        "42": [0, 0.75, 0.14917, 0, 0.51111],
-        "43": [0.05667, 0.56167, 0.03694, 0, 0.76666],
-        "44": [0.19444, 0.10556, 0, 0, 0.30667],
-        "45": [0, 0.43056, 0.02826, 0, 0.35778],
-        "46": [0, 0.10556, 0, 0, 0.30667],
-        "47": [0.25, 0.75, 0.16194, 0, 0.51111],
-        "48": [0, 0.64444, 0.13556, 0, 0.51111],
-        "49": [0, 0.64444, 0.13556, 0, 0.51111],
-        "50": [0, 0.64444, 0.13556, 0, 0.51111],
-        "51": [0, 0.64444, 0.13556, 0, 0.51111],
-        "52": [0.19444, 0.64444, 0.13556, 0, 0.51111],
-        "53": [0, 0.64444, 0.13556, 0, 0.51111],
-        "54": [0, 0.64444, 0.13556, 0, 0.51111],
-        "55": [0.19444, 0.64444, 0.13556, 0, 0.51111],
-        "56": [0, 0.64444, 0.13556, 0, 0.51111],
-        "57": [0, 0.64444, 0.13556, 0, 0.51111],
-        "58": [0, 0.43056, 0.0582, 0, 0.30667],
-        "59": [0.19444, 0.43056, 0.0582, 0, 0.30667],
-        "61": [-0.13313, 0.36687, 0.06616, 0, 0.76666],
-        "63": [0, 0.69444, 0.1225, 0, 0.51111],
-        "64": [0, 0.69444, 0.09597, 0, 0.76666],
-        "65": [0, 0.68333, 0, 0, 0.74333],
-        "66": [0, 0.68333, 0.10257, 0, 0.70389],
-        "67": [0, 0.68333, 0.14528, 0, 0.71555],
-        "68": [0, 0.68333, 0.09403, 0, 0.755],
-        "69": [0, 0.68333, 0.12028, 0, 0.67833],
-        "70": [0, 0.68333, 0.13305, 0, 0.65277],
-        "71": [0, 0.68333, 0.08722, 0, 0.77361],
-        "72": [0, 0.68333, 0.16389, 0, 0.74333],
-        "73": [0, 0.68333, 0.15806, 0, 0.38555],
-        "74": [0, 0.68333, 0.14028, 0, 0.525],
-        "75": [0, 0.68333, 0.14528, 0, 0.76888],
-        "76": [0, 0.68333, 0, 0, 0.62722],
-        "77": [0, 0.68333, 0.16389, 0, 0.89666],
-        "78": [0, 0.68333, 0.16389, 0, 0.74333],
-        "79": [0, 0.68333, 0.09403, 0, 0.76666],
-        "80": [0, 0.68333, 0.10257, 0, 0.67833],
-        "81": [0.19444, 0.68333, 0.09403, 0, 0.76666],
-        "82": [0, 0.68333, 0.03868, 0, 0.72944],
-        "83": [0, 0.68333, 0.11972, 0, 0.56222],
-        "84": [0, 0.68333, 0.13305, 0, 0.71555],
-        "85": [0, 0.68333, 0.16389, 0, 0.74333],
-        "86": [0, 0.68333, 0.18361, 0, 0.74333],
-        "87": [0, 0.68333, 0.18361, 0, 0.99888],
-        "88": [0, 0.68333, 0.15806, 0, 0.74333],
-        "89": [0, 0.68333, 0.19383, 0, 0.74333],
-        "90": [0, 0.68333, 0.14528, 0, 0.61333],
-        "91": [0.25, 0.75, 0.1875, 0, 0.30667],
-        "93": [0.25, 0.75, 0.10528, 0, 0.30667],
-        "94": [0, 0.69444, 0.06646, 0, 0.51111],
-        "95": [0.31, 0.12056, 0.09208, 0, 0.51111],
-        "97": [0, 0.43056, 0.07671, 0, 0.51111],
-        "98": [0, 0.69444, 0.06312, 0, 0.46],
-        "99": [0, 0.43056, 0.05653, 0, 0.46],
-        "100": [0, 0.69444, 0.10333, 0, 0.51111],
-        "101": [0, 0.43056, 0.07514, 0, 0.46],
-        "102": [0.19444, 0.69444, 0.21194, 0, 0.30667],
-        "103": [0.19444, 0.43056, 0.08847, 0, 0.46],
-        "104": [0, 0.69444, 0.07671, 0, 0.51111],
-        "105": [0, 0.65536, 0.1019, 0, 0.30667],
-        "106": [0.19444, 0.65536, 0.14467, 0, 0.30667],
-        "107": [0, 0.69444, 0.10764, 0, 0.46],
-        "108": [0, 0.69444, 0.10333, 0, 0.25555],
-        "109": [0, 0.43056, 0.07671, 0, 0.81777],
-        "110": [0, 0.43056, 0.07671, 0, 0.56222],
-        "111": [0, 0.43056, 0.06312, 0, 0.51111],
-        "112": [0.19444, 0.43056, 0.06312, 0, 0.51111],
-        "113": [0.19444, 0.43056, 0.08847, 0, 0.46],
-        "114": [0, 0.43056, 0.10764, 0, 0.42166],
-        "115": [0, 0.43056, 0.08208, 0, 0.40889],
-        "116": [0, 0.61508, 0.09486, 0, 0.33222],
-        "117": [0, 0.43056, 0.07671, 0, 0.53666],
-        "118": [0, 0.43056, 0.10764, 0, 0.46],
-        "119": [0, 0.43056, 0.10764, 0, 0.66444],
-        "120": [0, 0.43056, 0.12042, 0, 0.46389],
-        "121": [0.19444, 0.43056, 0.08847, 0, 0.48555],
-        "122": [0, 0.43056, 0.12292, 0, 0.40889],
-        "126": [0.35, 0.31786, 0.11585, 0, 0.51111],
-        "163": [0, 0.69444, 0, 0, 0.76909],
-        "168": [0, 0.66786, 0.10474, 0, 0.51111],
-        "176": [0, 0.69444, 0, 0, 0.83129],
-        "198": [0, 0.68333, 0.12028, 0, 0.88277],
-        "216": [0.04861, 0.73194, 0.09403, 0, 0.76666],
-        "223": [0.19444, 0.69444, 0.10514, 0, 0.53666],
-        "230": [0, 0.43056, 0.07514, 0, 0.71555],
-        "248": [0.09722, 0.52778, 0.09194, 0, 0.51111],
-        "305": [0, 0.43056, 0, 0.02778, 0.32246],
-        "338": [0, 0.68333, 0.12028, 0, 0.98499],
-        "339": [0, 0.43056, 0.07514, 0, 0.71555],
-        "567": [0.19444, 0.43056, 0, 0.08334, 0.38403],
-        "710": [0, 0.69444, 0.06646, 0, 0.51111],
-        "711": [0, 0.62847, 0.08295, 0, 0.51111],
-        "713": [0, 0.56167, 0.10333, 0, 0.51111],
-        "714": [0, 0.69444, 0.09694, 0, 0.51111],
-        "715": [0, 0.69444, 0, 0, 0.51111],
-        "728": [0, 0.69444, 0.10806, 0, 0.51111],
-        "729": [0, 0.66786, 0.11752, 0, 0.30667],
-        "730": [0, 0.69444, 0, 0, 0.83129],
-        "732": [0, 0.66786, 0.11585, 0, 0.51111],
-        "733": [0, 0.69444, 0.1225, 0, 0.51111],
-        "915": [0, 0.68333, 0.13305, 0, 0.62722],
-        "916": [0, 0.68333, 0, 0, 0.81777],
-        "920": [0, 0.68333, 0.09403, 0, 0.76666],
-        "923": [0, 0.68333, 0, 0, 0.69222],
-        "926": [0, 0.68333, 0.15294, 0, 0.66444],
-        "928": [0, 0.68333, 0.16389, 0, 0.74333],
-        "931": [0, 0.68333, 0.12028, 0, 0.71555],
-        "933": [0, 0.68333, 0.11111, 0, 0.76666],
-        "934": [0, 0.68333, 0.05986, 0, 0.71555],
-        "936": [0, 0.68333, 0.11111, 0, 0.76666],
-        "937": [0, 0.68333, 0.10257, 0, 0.71555],
-        "8211": [0, 0.43056, 0.09208, 0, 0.51111],
-        "8212": [0, 0.43056, 0.09208, 0, 1.02222],
-        "8216": [0, 0.69444, 0.12417, 0, 0.30667],
-        "8217": [0, 0.69444, 0.12417, 0, 0.30667],
-        "8220": [0, 0.69444, 0.1685, 0, 0.51444],
-        "8221": [0, 0.69444, 0.06961, 0, 0.51444],
-        "8463": [0, 0.68889, 0, 0, 0.54028],
-    },
-    "Main-Regular": {
-        "32": [0, 0, 0, 0, 0],
-        "33": [0, 0.69444, 0, 0, 0.27778],
-        "34": [0, 0.69444, 0, 0, 0.5],
-        "35": [0.19444, 0.69444, 0, 0, 0.83334],
-        "36": [0.05556, 0.75, 0, 0, 0.5],
-        "37": [0.05556, 0.75, 0, 0, 0.83334],
-        "38": [0, 0.69444, 0, 0, 0.77778],
-        "39": [0, 0.69444, 0, 0, 0.27778],
-        "40": [0.25, 0.75, 0, 0, 0.38889],
-        "41": [0.25, 0.75, 0, 0, 0.38889],
-        "42": [0, 0.75, 0, 0, 0.5],
-        "43": [0.08333, 0.58333, 0, 0, 0.77778],
-        "44": [0.19444, 0.10556, 0, 0, 0.27778],
-        "45": [0, 0.43056, 0, 0, 0.33333],
-        "46": [0, 0.10556, 0, 0, 0.27778],
-        "47": [0.25, 0.75, 0, 0, 0.5],
-        "48": [0, 0.64444, 0, 0, 0.5],
-        "49": [0, 0.64444, 0, 0, 0.5],
-        "50": [0, 0.64444, 0, 0, 0.5],
-        "51": [0, 0.64444, 0, 0, 0.5],
-        "52": [0, 0.64444, 0, 0, 0.5],
-        "53": [0, 0.64444, 0, 0, 0.5],
-        "54": [0, 0.64444, 0, 0, 0.5],
-        "55": [0, 0.64444, 0, 0, 0.5],
-        "56": [0, 0.64444, 0, 0, 0.5],
-        "57": [0, 0.64444, 0, 0, 0.5],
-        "58": [0, 0.43056, 0, 0, 0.27778],
-        "59": [0.19444, 0.43056, 0, 0, 0.27778],
-        "60": [0.0391, 0.5391, 0, 0, 0.77778],
-        "61": [-0.13313, 0.36687, 0, 0, 0.77778],
-        "62": [0.0391, 0.5391, 0, 0, 0.77778],
-        "63": [0, 0.69444, 0, 0, 0.47222],
-        "64": [0, 0.69444, 0, 0, 0.77778],
-        "65": [0, 0.68333, 0, 0, 0.75],
-        "66": [0, 0.68333, 0, 0, 0.70834],
-        "67": [0, 0.68333, 0, 0, 0.72222],
-        "68": [0, 0.68333, 0, 0, 0.76389],
-        "69": [0, 0.68333, 0, 0, 0.68056],
-        "70": [0, 0.68333, 0, 0, 0.65278],
-        "71": [0, 0.68333, 0, 0, 0.78472],
-        "72": [0, 0.68333, 0, 0, 0.75],
-        "73": [0, 0.68333, 0, 0, 0.36111],
-        "74": [0, 0.68333, 0, 0, 0.51389],
-        "75": [0, 0.68333, 0, 0, 0.77778],
-        "76": [0, 0.68333, 0, 0, 0.625],
-        "77": [0, 0.68333, 0, 0, 0.91667],
-        "78": [0, 0.68333, 0, 0, 0.75],
-        "79": [0, 0.68333, 0, 0, 0.77778],
-        "80": [0, 0.68333, 0, 0, 0.68056],
-        "81": [0.19444, 0.68333, 0, 0, 0.77778],
-        "82": [0, 0.68333, 0, 0, 0.73611],
-        "83": [0, 0.68333, 0, 0, 0.55556],
-        "84": [0, 0.68333, 0, 0, 0.72222],
-        "85": [0, 0.68333, 0, 0, 0.75],
-        "86": [0, 0.68333, 0.01389, 0, 0.75],
-        "87": [0, 0.68333, 0.01389, 0, 1.02778],
-        "88": [0, 0.68333, 0, 0, 0.75],
-        "89": [0, 0.68333, 0.025, 0, 0.75],
-        "90": [0, 0.68333, 0, 0, 0.61111],
-        "91": [0.25, 0.75, 0, 0, 0.27778],
-        "92": [0.25, 0.75, 0, 0, 0.5],
-        "93": [0.25, 0.75, 0, 0, 0.27778],
-        "94": [0, 0.69444, 0, 0, 0.5],
-        "95": [0.31, 0.12056, 0.02778, 0, 0.5],
-        "97": [0, 0.43056, 0, 0, 0.5],
-        "98": [0, 0.69444, 0, 0, 0.55556],
-        "99": [0, 0.43056, 0, 0, 0.44445],
-        "100": [0, 0.69444, 0, 0, 0.55556],
-        "101": [0, 0.43056, 0, 0, 0.44445],
-        "102": [0, 0.69444, 0.07778, 0, 0.30556],
-        "103": [0.19444, 0.43056, 0.01389, 0, 0.5],
-        "104": [0, 0.69444, 0, 0, 0.55556],
-        "105": [0, 0.66786, 0, 0, 0.27778],
-        "106": [0.19444, 0.66786, 0, 0, 0.30556],
-        "107": [0, 0.69444, 0, 0, 0.52778],
-        "108": [0, 0.69444, 0, 0, 0.27778],
-        "109": [0, 0.43056, 0, 0, 0.83334],
-        "110": [0, 0.43056, 0, 0, 0.55556],
-        "111": [0, 0.43056, 0, 0, 0.5],
-        "112": [0.19444, 0.43056, 0, 0, 0.55556],
-        "113": [0.19444, 0.43056, 0, 0, 0.52778],
-        "114": [0, 0.43056, 0, 0, 0.39167],
-        "115": [0, 0.43056, 0, 0, 0.39445],
-        "116": [0, 0.61508, 0, 0, 0.38889],
-        "117": [0, 0.43056, 0, 0, 0.55556],
-        "118": [0, 0.43056, 0.01389, 0, 0.52778],
-        "119": [0, 0.43056, 0.01389, 0, 0.72222],
-        "120": [0, 0.43056, 0, 0, 0.52778],
-        "121": [0.19444, 0.43056, 0.01389, 0, 0.52778],
-        "122": [0, 0.43056, 0, 0, 0.44445],
-        "123": [0.25, 0.75, 0, 0, 0.5],
-        "124": [0.25, 0.75, 0, 0, 0.27778],
-        "125": [0.25, 0.75, 0, 0, 0.5],
-        "126": [0.35, 0.31786, 0, 0, 0.5],
-        "160": [0, 0, 0, 0, 0],
-        "168": [0, 0.66786, 0, 0, 0.5],
-        "172": [0, 0.43056, 0, 0, 0.66667],
-        "176": [0, 0.69444, 0, 0, 0.75],
-        "177": [0.08333, 0.58333, 0, 0, 0.77778],
-        "198": [0, 0.68333, 0, 0, 0.90278],
-        "215": [0.08333, 0.58333, 0, 0, 0.77778],
-        "216": [0.04861, 0.73194, 0, 0, 0.77778],
-        "223": [0, 0.69444, 0, 0, 0.5],
-        "230": [0, 0.43056, 0, 0, 0.72222],
-        "247": [0.08333, 0.58333, 0, 0, 0.77778],
-        "248": [0.09722, 0.52778, 0, 0, 0.5],
-        "305": [0, 0.43056, 0, 0, 0.27778],
-        "338": [0, 0.68333, 0, 0, 1.01389],
-        "339": [0, 0.43056, 0, 0, 0.77778],
-        "567": [0.19444, 0.43056, 0, 0, 0.30556],
-        "710": [0, 0.69444, 0, 0, 0.5],
-        "711": [0, 0.62847, 0, 0, 0.5],
-        "713": [0, 0.56778, 0, 0, 0.5],
-        "714": [0, 0.69444, 0, 0, 0.5],
-        "715": [0, 0.69444, 0, 0, 0.5],
-        "728": [0, 0.69444, 0, 0, 0.5],
-        "729": [0, 0.66786, 0, 0, 0.27778],
-        "730": [0, 0.69444, 0, 0, 0.75],
-        "732": [0, 0.66786, 0, 0, 0.5],
-        "733": [0, 0.69444, 0, 0, 0.5],
-        "824": [0.19444, 0.69444, 0, 0, 0],
-        "915": [0, 0.68333, 0, 0, 0.625],
-        "916": [0, 0.68333, 0, 0, 0.83334],
-        "920": [0, 0.68333, 0, 0, 0.77778],
-        "923": [0, 0.68333, 0, 0, 0.69445],
-        "926": [0, 0.68333, 0, 0, 0.66667],
-        "928": [0, 0.68333, 0, 0, 0.75],
-        "931": [0, 0.68333, 0, 0, 0.72222],
-        "933": [0, 0.68333, 0, 0, 0.77778],
-        "934": [0, 0.68333, 0, 0, 0.72222],
-        "936": [0, 0.68333, 0, 0, 0.77778],
-        "937": [0, 0.68333, 0, 0, 0.72222],
-        "8211": [0, 0.43056, 0.02778, 0, 0.5],
-        "8212": [0, 0.43056, 0.02778, 0, 1.0],
-        "8216": [0, 0.69444, 0, 0, 0.27778],
-        "8217": [0, 0.69444, 0, 0, 0.27778],
-        "8220": [0, 0.69444, 0, 0, 0.5],
-        "8221": [0, 0.69444, 0, 0, 0.5],
-        "8224": [0.19444, 0.69444, 0, 0, 0.44445],
-        "8225": [0.19444, 0.69444, 0, 0, 0.44445],
-        "8230": [0, 0.12, 0, 0, 1015],
-        "8242": [0, 0.55556, 0, 0, 0.275],
-        "8407": [0, 0.71444, 0.15382, 0, 0.5],
-        "8463": [0, 0.68889, 0, 0, 0.54028],
-        "8465": [0, 0.69444, 0, 0, 0.72222],
-        "8467": [0, 0.69444, 0, 0.11111, 0.41667],
-        "8472": [0.19444, 0.43056, 0, 0.11111, 0.63646],
-        "8476": [0, 0.69444, 0, 0, 0.72222],
-        "8501": [0, 0.69444, 0, 0, 0.61111],
-        "8592": [-0.13313, 0.36687, 0, 0, 1.0],
-        "8593": [0.19444, 0.69444, 0, 0, 0.5],
-        "8594": [-0.13313, 0.36687, 0, 0, 1.0],
-        "8595": [0.19444, 0.69444, 0, 0, 0.5],
-        "8596": [-0.13313, 0.36687, 0, 0, 1.0],
-        "8597": [0.25, 0.75, 0, 0, 0.5],
-        "8598": [0.19444, 0.69444, 0, 0, 1.0],
-        "8599": [0.19444, 0.69444, 0, 0, 1.0],
-        "8600": [0.19444, 0.69444, 0, 0, 1.0],
-        "8601": [0.19444, 0.69444, 0, 0, 1.0],
-        "8614": [0.011, 0.511, 0, 0, 889],
-        "8617": [0.011, 0.511, 0, 0, 1015],
-        "8618": [0.011, 0.511, 0, 0, 1015],
-        "8636": [-0.13313, 0.36687, 0, 0, 1.0],
-        "8637": [-0.13313, 0.36687, 0, 0, 1.0],
-        "8640": [-0.13313, 0.36687, 0, 0, 1.0],
-        "8641": [-0.13313, 0.36687, 0, 0, 1.0],
-        "8652": [0.011, 0.671, 0, 0, 889],
-        "8656": [-0.13313, 0.36687, 0, 0, 1.0],
-        "8657": [0.19444, 0.69444, 0, 0, 0.61111],
-        "8658": [-0.13313, 0.36687, 0, 0, 1.0],
-        "8659": [0.19444, 0.69444, 0, 0, 0.61111],
-        "8660": [-0.13313, 0.36687, 0, 0, 1.0],
-        "8661": [0.25, 0.75, 0, 0, 0.61111],
-        "8704": [0, 0.69444, 0, 0, 0.55556],
-        "8706": [0, 0.69444, 0.05556, 0.08334, 0.5309],
-        "8707": [0, 0.69444, 0, 0, 0.55556],
-        "8709": [0.05556, 0.75, 0, 0, 0.5],
-        "8711": [0, 0.68333, 0, 0, 0.83334],
-        "8712": [0.0391, 0.5391, 0, 0, 0.66667],
-        "8715": [0.0391, 0.5391, 0, 0, 0.66667],
-        "8722": [0.08333, 0.58333, 0, 0, 0.77778],
-        "8723": [0.08333, 0.58333, 0, 0, 0.77778],
-        "8725": [0.25, 0.75, 0, 0, 0.5],
-        "8726": [0.25, 0.75, 0, 0, 0.5],
-        "8727": [-0.03472, 0.46528, 0, 0, 0.5],
-        "8728": [-0.05555, 0.44445, 0, 0, 0.5],
-        "8729": [-0.05555, 0.44445, 0, 0, 0.5],
-        "8730": [0.2, 0.8, 0, 0, 0.83334],
-        "8733": [0, 0.43056, 0, 0, 0.77778],
-        "8734": [0, 0.43056, 0, 0, 1.0],
-        "8736": [0, 0.69224, 0, 0, 0.72222],
-        "8739": [0.25, 0.75, 0, 0, 0.27778],
-        "8741": [0.25, 0.75, 0, 0, 0.5],
-        "8743": [0, 0.55556, 0, 0, 0.66667],
-        "8744": [0, 0.55556, 0, 0, 0.66667],
-        "8745": [0, 0.55556, 0, 0, 0.66667],
-        "8746": [0, 0.55556, 0, 0, 0.66667],
-        "8747": [0.19444, 0.69444, 0.11111, 0, 0.41667],
-        "8764": [-0.13313, 0.36687, 0, 0, 0.77778],
-        "8768": [0.19444, 0.69444, 0, 0, 0.27778],
-        "8771": [-0.03625, 0.46375, 0, 0, 0.77778],
-        "8773": [-0.022, 0.589, 0, 0, 667],
-        "8776": [-0.01688, 0.48312, 0, 0, 0.77778],
-        "8781": [-0.03625, 0.46375, 0, 0, 0.77778],
-        "8784": [-0.133, 0.67, 0, 0, 666],
-        "8800": [0.215, 0.716, 0, 0, 666],
-        "8801": [-0.03625, 0.46375, 0, 0, 0.77778],
-        "8804": [0.13597, 0.63597, 0, 0, 0.77778],
-        "8805": [0.13597, 0.63597, 0, 0, 0.77778],
-        "8810": [0.0391, 0.5391, 0, 0, 1.0],
-        "8811": [0.0391, 0.5391, 0, 0, 1.0],
-        "8826": [0.0391, 0.5391, 0, 0, 0.77778],
-        "8827": [0.0391, 0.5391, 0, 0, 0.77778],
-        "8834": [0.0391, 0.5391, 0, 0, 0.77778],
-        "8835": [0.0391, 0.5391, 0, 0, 0.77778],
-        "8838": [0.13597, 0.63597, 0, 0, 0.77778],
-        "8839": [0.13597, 0.63597, 0, 0, 0.77778],
-        "8846": [0, 0.55556, 0, 0, 0.66667],
-        "8849": [0.13597, 0.63597, 0, 0, 0.77778],
-        "8850": [0.13597, 0.63597, 0, 0, 0.77778],
-        "8851": [0, 0.55556, 0, 0, 0.66667],
-        "8852": [0, 0.55556, 0, 0, 0.66667],
-        "8853": [0.08333, 0.58333, 0, 0, 0.77778],
-        "8854": [0.08333, 0.58333, 0, 0, 0.77778],
-        "8855": [0.08333, 0.58333, 0, 0, 0.77778],
-        "8856": [0.08333, 0.58333, 0, 0, 0.77778],
-        "8857": [0.08333, 0.58333, 0, 0, 0.77778],
-        "8866": [0, 0.69444, 0, 0, 0.61111],
-        "8867": [0, 0.69444, 0, 0, 0.61111],
-        "8868": [0, 0.69444, 0, 0, 0.77778],
-        "8869": [0, 0.69444, 0, 0, 0.77778],
-        "8872": [0.249, 0.75, 0, 0, 692],
-        "8900": [-0.05555, 0.44445, 0, 0, 0.5],
-        "8901": [-0.05555, 0.44445, 0, 0, 0.27778],
-        "8902": [-0.03472, 0.46528, 0, 0, 0.5],
-        "8904": [0.005, 0.505, 0, 0, 847],
-        "8942": [0.03, 0.9, 0, 0, 121],
-        "8943": [-0.19, 0.31, 0, 0, 1015],
-        "8945": [-0.1, 0.82, 0, 0, 1015],
-        "8968": [0.25, 0.75, 0, 0, 0.44445],
-        "8969": [0.25, 0.75, 0, 0, 0.44445],
-        "8970": [0.25, 0.75, 0, 0, 0.44445],
-        "8971": [0.25, 0.75, 0, 0, 0.44445],
-        "8994": [-0.14236, 0.35764, 0, 0, 1.0],
-        "8995": [-0.14236, 0.35764, 0, 0, 1.0],
-        "9136": [0.244, 0.744, 0, 0, 301],
-        "9137": [0.244, 0.744, 0, 0, 301],
-        "9651": [0.19444, 0.69444, 0, 0, 0.88889],
-        "9657": [-0.03472, 0.46528, 0, 0, 0.5],
-        "9661": [0.19444, 0.69444, 0, 0, 0.88889],
-        "9667": [-0.03472, 0.46528, 0, 0, 0.5],
-        "9711": [0.19444, 0.69444, 0, 0, 1.0],
-        "9824": [0.12963, 0.69444, 0, 0, 0.77778],
-        "9825": [0.12963, 0.69444, 0, 0, 0.77778],
-        "9826": [0.12963, 0.69444, 0, 0, 0.77778],
-        "9827": [0.12963, 0.69444, 0, 0, 0.77778],
-        "9837": [0, 0.75, 0, 0, 0.38889],
-        "9838": [0.19444, 0.69444, 0, 0, 0.38889],
-        "9839": [0.19444, 0.69444, 0, 0, 0.38889],
-        "10216": [0.25, 0.75, 0, 0, 0.38889],
-        "10217": [0.25, 0.75, 0, 0, 0.38889],
-        "10222": [0.244, 0.744, 0, 0, 184],
-        "10223": [0.244, 0.744, 0, 0, 184],
-        "10229": [0.011, 0.511, 0, 0, 1470],
-        "10230": [0.011, 0.511, 0, 0, 1469],
-        "10231": [0.011, 0.511, 0, 0, 1748],
-        "10232": [0.024, 0.525, 0, 0, 1497],
-        "10233": [0.024, 0.525, 0, 0, 1526],
-        "10234": [0.024, 0.525, 0, 0, 1746],
-        "10236": [0.011, 0.511, 0, 0, 1498],
-        "10815": [0, 0.68333, 0, 0, 0.75],
-        "10927": [0.13597, 0.63597, 0, 0, 0.77778],
-        "10928": [0.13597, 0.63597, 0, 0, 0.77778],
-    },
-    "Math-BoldItalic": {
-        "47": [0.19444, 0.69444, 0, 0, 0],
-        "65": [0, 0.68611, 0, 0, 0.86944],
-        "66": [0, 0.68611, 0.04835, 0, 0.8664],
-        "67": [0, 0.68611, 0.06979, 0, 0.81694],
-        "68": [0, 0.68611, 0.03194, 0, 0.93812],
-        "69": [0, 0.68611, 0.05451, 0, 0.81007],
-        "70": [0, 0.68611, 0.15972, 0, 0.68889],
-        "71": [0, 0.68611, 0, 0, 0.88673],
-        "72": [0, 0.68611, 0.08229, 0, 0.98229],
-        "73": [0, 0.68611, 0.07778, 0, 0.51111],
-        "74": [0, 0.68611, 0.10069, 0, 0.63125],
-        "75": [0, 0.68611, 0.06979, 0, 0.97118],
-        "76": [0, 0.68611, 0, 0, 0.75555],
-        "77": [0, 0.68611, 0.11424, 0, 1.14201],
-        "78": [0, 0.68611, 0.11424, 0, 0.95034],
-        "79": [0, 0.68611, 0.03194, 0, 0.83666],
-        "80": [0, 0.68611, 0.15972, 0, 0.72309],
-        "81": [0.19444, 0.68611, 0, 0, 0.86861],
-        "82": [0, 0.68611, 0.00421, 0, 0.87235],
-        "83": [0, 0.68611, 0.05382, 0, 0.69271],
-        "84": [0, 0.68611, 0.15972, 0, 0.63663],
-        "85": [0, 0.68611, 0.11424, 0, 0.80027],
-        "86": [0, 0.68611, 0.25555, 0, 0.67778],
-        "87": [0, 0.68611, 0.15972, 0, 1.09305],
-        "88": [0, 0.68611, 0.07778, 0, 0.94722],
-        "89": [0, 0.68611, 0.25555, 0, 0.67458],
-        "90": [0, 0.68611, 0.06979, 0, 0.77257],
-        "97": [0, 0.44444, 0, 0, 0.63287],
-        "98": [0, 0.69444, 0, 0, 0.52083],
-        "99": [0, 0.44444, 0, 0, 0.51342],
-        "100": [0, 0.69444, 0, 0, 0.60972],
-        "101": [0, 0.44444, 0, 0, 0.55361],
-        "102": [0.19444, 0.69444, 0.11042, 0, 0.56806],
-        "103": [0.19444, 0.44444, 0.03704, 0, 0.5449],
-        "104": [0, 0.69444, 0, 0, 0.66759],
-        "105": [0, 0.69326, 0, 0, 0.4048],
-        "106": [0.19444, 0.69326, 0.0622, 0, 0.47083],
-        "107": [0, 0.69444, 0.01852, 0, 0.6037],
-        "108": [0, 0.69444, 0.0088, 0, 0.34815],
-        "109": [0, 0.44444, 0, 0, 1.0324],
-        "110": [0, 0.44444, 0, 0, 0.71296],
-        "111": [0, 0.44444, 0, 0, 0.58472],
-        "112": [0.19444, 0.44444, 0, 0, 0.60092],
-        "113": [0.19444, 0.44444, 0.03704, 0, 0.54213],
-        "114": [0, 0.44444, 0.03194, 0, 0.5287],
-        "115": [0, 0.44444, 0, 0, 0.53125],
-        "116": [0, 0.63492, 0, 0, 0.41528],
-        "117": [0, 0.44444, 0, 0, 0.68102],
-        "118": [0, 0.44444, 0.03704, 0, 0.56666],
-        "119": [0, 0.44444, 0.02778, 0, 0.83148],
-        "120": [0, 0.44444, 0, 0, 0.65903],
-        "121": [0.19444, 0.44444, 0.03704, 0, 0.59028],
-        "122": [0, 0.44444, 0.04213, 0, 0.55509],
-        "915": [0, 0.68611, 0.15972, 0, 0.65694],
-        "916": [0, 0.68611, 0, 0, 0.95833],
-        "920": [0, 0.68611, 0.03194, 0, 0.86722],
-        "923": [0, 0.68611, 0, 0, 0.80555],
-        "926": [0, 0.68611, 0.07458, 0, 0.84125],
-        "928": [0, 0.68611, 0.08229, 0, 0.98229],
-        "931": [0, 0.68611, 0.05451, 0, 0.88507],
-        "933": [0, 0.68611, 0.15972, 0, 0.67083],
-        "934": [0, 0.68611, 0, 0, 0.76666],
-        "936": [0, 0.68611, 0.11653, 0, 0.71402],
-        "937": [0, 0.68611, 0.04835, 0, 0.8789],
-        "945": [0, 0.44444, 0, 0, 0.76064],
-        "946": [0.19444, 0.69444, 0.03403, 0, 0.65972],
-        "947": [0.19444, 0.44444, 0.06389, 0, 0.59003],
-        "948": [0, 0.69444, 0.03819, 0, 0.52222],
-        "949": [0, 0.44444, 0, 0, 0.52882],
-        "950": [0.19444, 0.69444, 0.06215, 0, 0.50833],
-        "951": [0.19444, 0.44444, 0.03704, 0, 0.6],
-        "952": [0, 0.69444, 0.03194, 0, 0.5618],
-        "953": [0, 0.44444, 0, 0, 0.41204],
-        "954": [0, 0.44444, 0, 0, 0.66759],
-        "955": [0, 0.69444, 0, 0, 0.67083],
-        "956": [0.19444, 0.44444, 0, 0, 0.70787],
-        "957": [0, 0.44444, 0.06898, 0, 0.57685],
-        "958": [0.19444, 0.69444, 0.03021, 0, 0.50833],
-        "959": [0, 0.44444, 0, 0, 0.58472],
-        "960": [0, 0.44444, 0.03704, 0, 0.68241],
-        "961": [0.19444, 0.44444, 0, 0, 0.6118],
-        "962": [0.09722, 0.44444, 0.07917, 0, 0.42361],
-        "963": [0, 0.44444, 0.03704, 0, 0.68588],
-        "964": [0, 0.44444, 0.13472, 0, 0.52083],
-        "965": [0, 0.44444, 0.03704, 0, 0.63055],
-        "966": [0.19444, 0.44444, 0, 0, 0.74722],
-        "967": [0.19444, 0.44444, 0, 0, 0.71805],
-        "968": [0.19444, 0.69444, 0.03704, 0, 0.75833],
-        "969": [0, 0.44444, 0.03704, 0, 0.71782],
-        "977": [0, 0.69444, 0, 0, 0.69155],
-        "981": [0.19444, 0.69444, 0, 0, 0.7125],
-        "982": [0, 0.44444, 0.03194, 0, 0.975],
-        "1009": [0.19444, 0.44444, 0, 0, 0.6118],
-        "1013": [0, 0.44444, 0, 0, 0.48333],
-    },
-    "Math-Italic": {
-        "47": [0.19444, 0.69444, 0, 0, 0],
-        "65": [0, 0.68333, 0, 0.13889, 0.75],
-        "66": [0, 0.68333, 0.05017, 0.08334, 0.75851],
-        "67": [0, 0.68333, 0.07153, 0.08334, 0.71472],
-        "68": [0, 0.68333, 0.02778, 0.05556, 0.82792],
-        "69": [0, 0.68333, 0.05764, 0.08334, 0.7382],
-        "70": [0, 0.68333, 0.13889, 0.08334, 0.64306],
-        "71": [0, 0.68333, 0, 0.08334, 0.78625],
-        "72": [0, 0.68333, 0.08125, 0.05556, 0.83125],
-        "73": [0, 0.68333, 0.07847, 0.11111, 0.43958],
-        "74": [0, 0.68333, 0.09618, 0.16667, 0.55451],
-        "75": [0, 0.68333, 0.07153, 0.05556, 0.84931],
-        "76": [0, 0.68333, 0, 0.02778, 0.68056],
-        "77": [0, 0.68333, 0.10903, 0.08334, 0.97014],
-        "78": [0, 0.68333, 0.10903, 0.08334, 0.80347],
-        "79": [0, 0.68333, 0.02778, 0.08334, 0.76278],
-        "80": [0, 0.68333, 0.13889, 0.08334, 0.64201],
-        "81": [0.19444, 0.68333, 0, 0.08334, 0.79056],
-        "82": [0, 0.68333, 0.00773, 0.08334, 0.75929],
-        "83": [0, 0.68333, 0.05764, 0.08334, 0.6132],
-        "84": [0, 0.68333, 0.13889, 0.08334, 0.58438],
-        "85": [0, 0.68333, 0.10903, 0.02778, 0.68278],
-        "86": [0, 0.68333, 0.22222, 0, 0.58333],
-        "87": [0, 0.68333, 0.13889, 0, 0.94445],
-        "88": [0, 0.68333, 0.07847, 0.08334, 0.82847],
-        "89": [0, 0.68333, 0.22222, 0, 0.58056],
-        "90": [0, 0.68333, 0.07153, 0.08334, 0.68264],
-        "97": [0, 0.43056, 0, 0, 0.52859],
-        "98": [0, 0.69444, 0, 0, 0.42917],
-        "99": [0, 0.43056, 0, 0.05556, 0.43276],
-        "100": [0, 0.69444, 0, 0.16667, 0.52049],
-        "101": [0, 0.43056, 0, 0.05556, 0.46563],
-        "102": [0.19444, 0.69444, 0.10764, 0.16667, 0.48959],
-        "103": [0.19444, 0.43056, 0.03588, 0.02778, 0.47697],
-        "104": [0, 0.69444, 0, 0, 0.57616],
-        "105": [0, 0.65952, 0, 0, 0.34451],
-        "106": [0.19444, 0.65952, 0.05724, 0, 0.41181],
-        "107": [0, 0.69444, 0.03148, 0, 0.5206],
-        "108": [0, 0.69444, 0.01968, 0.08334, 0.29838],
-        "109": [0, 0.43056, 0, 0, 0.87801],
-        "110": [0, 0.43056, 0, 0, 0.60023],
-        "111": [0, 0.43056, 0, 0.05556, 0.48472],
-        "112": [0.19444, 0.43056, 0, 0.08334, 0.50313],
-        "113": [0.19444, 0.43056, 0.03588, 0.08334, 0.44641],
-        "114": [0, 0.43056, 0.02778, 0.05556, 0.45116],
-        "115": [0, 0.43056, 0, 0.05556, 0.46875],
-        "116": [0, 0.61508, 0, 0.08334, 0.36111],
-        "117": [0, 0.43056, 0, 0.02778, 0.57246],
-        "118": [0, 0.43056, 0.03588, 0.02778, 0.48472],
-        "119": [0, 0.43056, 0.02691, 0.08334, 0.71592],
-        "120": [0, 0.43056, 0, 0.02778, 0.57153],
-        "121": [0.19444, 0.43056, 0.03588, 0.05556, 0.49028],
-        "122": [0, 0.43056, 0.04398, 0.05556, 0.46505],
-        "915": [0, 0.68333, 0.13889, 0.08334, 0.61528],
-        "916": [0, 0.68333, 0, 0.16667, 0.83334],
-        "920": [0, 0.68333, 0.02778, 0.08334, 0.76278],
-        "923": [0, 0.68333, 0, 0.16667, 0.69445],
-        "926": [0, 0.68333, 0.07569, 0.08334, 0.74236],
-        "928": [0, 0.68333, 0.08125, 0.05556, 0.83125],
-        "931": [0, 0.68333, 0.05764, 0.08334, 0.77986],
-        "933": [0, 0.68333, 0.13889, 0.05556, 0.58333],
-        "934": [0, 0.68333, 0, 0.08334, 0.66667],
-        "936": [0, 0.68333, 0.11, 0.05556, 0.61222],
-        "937": [0, 0.68333, 0.05017, 0.08334, 0.7724],
-        "945": [0, 0.43056, 0.0037, 0.02778, 0.6397],
-        "946": [0.19444, 0.69444, 0.05278, 0.08334, 0.56563],
-        "947": [0.19444, 0.43056, 0.05556, 0, 0.51773],
-        "948": [0, 0.69444, 0.03785, 0.05556, 0.44444],
-        "949": [0, 0.43056, 0, 0.08334, 0.46632],
-        "950": [0.19444, 0.69444, 0.07378, 0.08334, 0.4375],
-        "951": [0.19444, 0.43056, 0.03588, 0.05556, 0.49653],
-        "952": [0, 0.69444, 0.02778, 0.08334, 0.46944],
-        "953": [0, 0.43056, 0, 0.05556, 0.35394],
-        "954": [0, 0.43056, 0, 0, 0.57616],
-        "955": [0, 0.69444, 0, 0, 0.58334],
-        "956": [0.19444, 0.43056, 0, 0.02778, 0.60255],
-        "957": [0, 0.43056, 0.06366, 0.02778, 0.49398],
-        "958": [0.19444, 0.69444, 0.04601, 0.11111, 0.4375],
-        "959": [0, 0.43056, 0, 0.05556, 0.48472],
-        "960": [0, 0.43056, 0.03588, 0, 0.57003],
-        "961": [0.19444, 0.43056, 0, 0.08334, 0.51702],
-        "962": [0.09722, 0.43056, 0.07986, 0.08334, 0.36285],
-        "963": [0, 0.43056, 0.03588, 0, 0.57141],
-        "964": [0, 0.43056, 0.1132, 0.02778, 0.43715],
-        "965": [0, 0.43056, 0.03588, 0.02778, 0.54028],
-        "966": [0.19444, 0.43056, 0, 0.08334, 0.65417],
-        "967": [0.19444, 0.43056, 0, 0.05556, 0.62569],
-        "968": [0.19444, 0.69444, 0.03588, 0.11111, 0.65139],
-        "969": [0, 0.43056, 0.03588, 0, 0.62245],
-        "977": [0, 0.69444, 0, 0.08334, 0.59144],
-        "981": [0.19444, 0.69444, 0, 0.08334, 0.59583],
-        "982": [0, 0.43056, 0.02778, 0, 0.82813],
-        "1009": [0.19444, 0.43056, 0, 0.08334, 0.51702],
-        "1013": [0, 0.43056, 0, 0.05556, 0.4059],
-    },
-    "Math-Regular": {
-        "65": [0, 0.68333, 0, 0.13889, 0.75],
-        "66": [0, 0.68333, 0.05017, 0.08334, 0.75851],
-        "67": [0, 0.68333, 0.07153, 0.08334, 0.71472],
-        "68": [0, 0.68333, 0.02778, 0.05556, 0.82792],
-        "69": [0, 0.68333, 0.05764, 0.08334, 0.7382],
-        "70": [0, 0.68333, 0.13889, 0.08334, 0.64306],
-        "71": [0, 0.68333, 0, 0.08334, 0.78625],
-        "72": [0, 0.68333, 0.08125, 0.05556, 0.83125],
-        "73": [0, 0.68333, 0.07847, 0.11111, 0.43958],
-        "74": [0, 0.68333, 0.09618, 0.16667, 0.55451],
-        "75": [0, 0.68333, 0.07153, 0.05556, 0.84931],
-        "76": [0, 0.68333, 0, 0.02778, 0.68056],
-        "77": [0, 0.68333, 0.10903, 0.08334, 0.97014],
-        "78": [0, 0.68333, 0.10903, 0.08334, 0.80347],
-        "79": [0, 0.68333, 0.02778, 0.08334, 0.76278],
-        "80": [0, 0.68333, 0.13889, 0.08334, 0.64201],
-        "81": [0.19444, 0.68333, 0, 0.08334, 0.79056],
-        "82": [0, 0.68333, 0.00773, 0.08334, 0.75929],
-        "83": [0, 0.68333, 0.05764, 0.08334, 0.6132],
-        "84": [0, 0.68333, 0.13889, 0.08334, 0.58438],
-        "85": [0, 0.68333, 0.10903, 0.02778, 0.68278],
-        "86": [0, 0.68333, 0.22222, 0, 0.58333],
-        "87": [0, 0.68333, 0.13889, 0, 0.94445],
-        "88": [0, 0.68333, 0.07847, 0.08334, 0.82847],
-        "89": [0, 0.68333, 0.22222, 0, 0.58056],
-        "90": [0, 0.68333, 0.07153, 0.08334, 0.68264],
-        "97": [0, 0.43056, 0, 0, 0.52859],
-        "98": [0, 0.69444, 0, 0, 0.42917],
-        "99": [0, 0.43056, 0, 0.05556, 0.43276],
-        "100": [0, 0.69444, 0, 0.16667, 0.52049],
-        "101": [0, 0.43056, 0, 0.05556, 0.46563],
-        "102": [0.19444, 0.69444, 0.10764, 0.16667, 0.48959],
-        "103": [0.19444, 0.43056, 0.03588, 0.02778, 0.47697],
-        "104": [0, 0.69444, 0, 0, 0.57616],
-        "105": [0, 0.65952, 0, 0, 0.34451],
-        "106": [0.19444, 0.65952, 0.05724, 0, 0.41181],
-        "107": [0, 0.69444, 0.03148, 0, 0.5206],
-        "108": [0, 0.69444, 0.01968, 0.08334, 0.29838],
-        "109": [0, 0.43056, 0, 0, 0.87801],
-        "110": [0, 0.43056, 0, 0, 0.60023],
-        "111": [0, 0.43056, 0, 0.05556, 0.48472],
-        "112": [0.19444, 0.43056, 0, 0.08334, 0.50313],
-        "113": [0.19444, 0.43056, 0.03588, 0.08334, 0.44641],
-        "114": [0, 0.43056, 0.02778, 0.05556, 0.45116],
-        "115": [0, 0.43056, 0, 0.05556, 0.46875],
-        "116": [0, 0.61508, 0, 0.08334, 0.36111],
-        "117": [0, 0.43056, 0, 0.02778, 0.57246],
-        "118": [0, 0.43056, 0.03588, 0.02778, 0.48472],
-        "119": [0, 0.43056, 0.02691, 0.08334, 0.71592],
-        "120": [0, 0.43056, 0, 0.02778, 0.57153],
-        "121": [0.19444, 0.43056, 0.03588, 0.05556, 0.49028],
-        "122": [0, 0.43056, 0.04398, 0.05556, 0.46505],
-        "915": [0, 0.68333, 0.13889, 0.08334, 0.61528],
-        "916": [0, 0.68333, 0, 0.16667, 0.83334],
-        "920": [0, 0.68333, 0.02778, 0.08334, 0.76278],
-        "923": [0, 0.68333, 0, 0.16667, 0.69445],
-        "926": [0, 0.68333, 0.07569, 0.08334, 0.74236],
-        "928": [0, 0.68333, 0.08125, 0.05556, 0.83125],
-        "931": [0, 0.68333, 0.05764, 0.08334, 0.77986],
-        "933": [0, 0.68333, 0.13889, 0.05556, 0.58333],
-        "934": [0, 0.68333, 0, 0.08334, 0.66667],
-        "936": [0, 0.68333, 0.11, 0.05556, 0.61222],
-        "937": [0, 0.68333, 0.05017, 0.08334, 0.7724],
-        "945": [0, 0.43056, 0.0037, 0.02778, 0.6397],
-        "946": [0.19444, 0.69444, 0.05278, 0.08334, 0.56563],
-        "947": [0.19444, 0.43056, 0.05556, 0, 0.51773],
-        "948": [0, 0.69444, 0.03785, 0.05556, 0.44444],
-        "949": [0, 0.43056, 0, 0.08334, 0.46632],
-        "950": [0.19444, 0.69444, 0.07378, 0.08334, 0.4375],
-        "951": [0.19444, 0.43056, 0.03588, 0.05556, 0.49653],
-        "952": [0, 0.69444, 0.02778, 0.08334, 0.46944],
-        "953": [0, 0.43056, 0, 0.05556, 0.35394],
-        "954": [0, 0.43056, 0, 0, 0.57616],
-        "955": [0, 0.69444, 0, 0, 0.58334],
-        "956": [0.19444, 0.43056, 0, 0.02778, 0.60255],
-        "957": [0, 0.43056, 0.06366, 0.02778, 0.49398],
-        "958": [0.19444, 0.69444, 0.04601, 0.11111, 0.4375],
-        "959": [0, 0.43056, 0, 0.05556, 0.48472],
-        "960": [0, 0.43056, 0.03588, 0, 0.57003],
-        "961": [0.19444, 0.43056, 0, 0.08334, 0.51702],
-        "962": [0.09722, 0.43056, 0.07986, 0.08334, 0.36285],
-        "963": [0, 0.43056, 0.03588, 0, 0.57141],
-        "964": [0, 0.43056, 0.1132, 0.02778, 0.43715],
-        "965": [0, 0.43056, 0.03588, 0.02778, 0.54028],
-        "966": [0.19444, 0.43056, 0, 0.08334, 0.65417],
-        "967": [0.19444, 0.43056, 0, 0.05556, 0.62569],
-        "968": [0.19444, 0.69444, 0.03588, 0.11111, 0.65139],
-        "969": [0, 0.43056, 0.03588, 0, 0.62245],
-        "977": [0, 0.69444, 0, 0.08334, 0.59144],
-        "981": [0.19444, 0.69444, 0, 0.08334, 0.59583],
-        "982": [0, 0.43056, 0.02778, 0, 0.82813],
-        "1009": [0.19444, 0.43056, 0, 0.08334, 0.51702],
-        "1013": [0, 0.43056, 0, 0.05556, 0.4059],
-    },
-    "SansSerif-Bold": {
-        "33": [0, 0.69444, 0, 0, 0.36667],
-        "34": [0, 0.69444, 0, 0, 0.55834],
-        "35": [0.19444, 0.69444, 0, 0, 0.91667],
-        "36": [0.05556, 0.75, 0, 0, 0.55],
-        "37": [0.05556, 0.75, 0, 0, 1.02912],
-        "38": [0, 0.69444, 0, 0, 0.83056],
-        "39": [0, 0.69444, 0, 0, 0.30556],
-        "40": [0.25, 0.75, 0, 0, 0.42778],
-        "41": [0.25, 0.75, 0, 0, 0.42778],
-        "42": [0, 0.75, 0, 0, 0.55],
-        "43": [0.11667, 0.61667, 0, 0, 0.85556],
-        "44": [0.10556, 0.13056, 0, 0, 0.30556],
-        "45": [0, 0.45833, 0, 0, 0.36667],
-        "46": [0, 0.13056, 0, 0, 0.30556],
-        "47": [0.25, 0.75, 0, 0, 0.55],
-        "48": [0, 0.69444, 0, 0, 0.55],
-        "49": [0, 0.69444, 0, 0, 0.55],
-        "50": [0, 0.69444, 0, 0, 0.55],
-        "51": [0, 0.69444, 0, 0, 0.55],
-        "52": [0, 0.69444, 0, 0, 0.55],
-        "53": [0, 0.69444, 0, 0, 0.55],
-        "54": [0, 0.69444, 0, 0, 0.55],
-        "55": [0, 0.69444, 0, 0, 0.55],
-        "56": [0, 0.69444, 0, 0, 0.55],
-        "57": [0, 0.69444, 0, 0, 0.55],
-        "58": [0, 0.45833, 0, 0, 0.30556],
-        "59": [0.10556, 0.45833, 0, 0, 0.30556],
-        "61": [-0.09375, 0.40625, 0, 0, 0.85556],
-        "63": [0, 0.69444, 0, 0, 0.51945],
-        "64": [0, 0.69444, 0, 0, 0.73334],
-        "65": [0, 0.69444, 0, 0, 0.73334],
-        "66": [0, 0.69444, 0, 0, 0.73334],
-        "67": [0, 0.69444, 0, 0, 0.70278],
-        "68": [0, 0.69444, 0, 0, 0.79445],
-        "69": [0, 0.69444, 0, 0, 0.64167],
-        "70": [0, 0.69444, 0, 0, 0.61111],
-        "71": [0, 0.69444, 0, 0, 0.73334],
-        "72": [0, 0.69444, 0, 0, 0.79445],
-        "73": [0, 0.69444, 0, 0, 0.33056],
-        "74": [0, 0.69444, 0, 0, 0.51945],
-        "75": [0, 0.69444, 0, 0, 0.76389],
-        "76": [0, 0.69444, 0, 0, 0.58056],
-        "77": [0, 0.69444, 0, 0, 0.97778],
-        "78": [0, 0.69444, 0, 0, 0.79445],
-        "79": [0, 0.69444, 0, 0, 0.79445],
-        "80": [0, 0.69444, 0, 0, 0.70278],
-        "81": [0.10556, 0.69444, 0, 0, 0.79445],
-        "82": [0, 0.69444, 0, 0, 0.70278],
-        "83": [0, 0.69444, 0, 0, 0.61111],
-        "84": [0, 0.69444, 0, 0, 0.73334],
-        "85": [0, 0.69444, 0, 0, 0.76389],
-        "86": [0, 0.69444, 0.01528, 0, 0.73334],
-        "87": [0, 0.69444, 0.01528, 0, 1.03889],
-        "88": [0, 0.69444, 0, 0, 0.73334],
-        "89": [0, 0.69444, 0.0275, 0, 0.73334],
-        "90": [0, 0.69444, 0, 0, 0.67223],
-        "91": [0.25, 0.75, 0, 0, 0.34306],
-        "93": [0.25, 0.75, 0, 0, 0.34306],
-        "94": [0, 0.69444, 0, 0, 0.55],
-        "95": [0.35, 0.10833, 0.03056, 0, 0.55],
-        "97": [0, 0.45833, 0, 0, 0.525],
-        "98": [0, 0.69444, 0, 0, 0.56111],
-        "99": [0, 0.45833, 0, 0, 0.48889],
-        "100": [0, 0.69444, 0, 0, 0.56111],
-        "101": [0, 0.45833, 0, 0, 0.51111],
-        "102": [0, 0.69444, 0.07639, 0, 0.33611],
-        "103": [0.19444, 0.45833, 0.01528, 0, 0.55],
-        "104": [0, 0.69444, 0, 0, 0.56111],
-        "105": [0, 0.69444, 0, 0, 0.25556],
-        "106": [0.19444, 0.69444, 0, 0, 0.28611],
-        "107": [0, 0.69444, 0, 0, 0.53056],
-        "108": [0, 0.69444, 0, 0, 0.25556],
-        "109": [0, 0.45833, 0, 0, 0.86667],
-        "110": [0, 0.45833, 0, 0, 0.56111],
-        "111": [0, 0.45833, 0, 0, 0.55],
-        "112": [0.19444, 0.45833, 0, 0, 0.56111],
-        "113": [0.19444, 0.45833, 0, 0, 0.56111],
-        "114": [0, 0.45833, 0.01528, 0, 0.37222],
-        "115": [0, 0.45833, 0, 0, 0.42167],
-        "116": [0, 0.58929, 0, 0, 0.40417],
-        "117": [0, 0.45833, 0, 0, 0.56111],
-        "118": [0, 0.45833, 0.01528, 0, 0.5],
-        "119": [0, 0.45833, 0.01528, 0, 0.74445],
-        "120": [0, 0.45833, 0, 0, 0.5],
-        "121": [0.19444, 0.45833, 0.01528, 0, 0.5],
-        "122": [0, 0.45833, 0, 0, 0.47639],
-        "126": [0.35, 0.34444, 0, 0, 0.55],
-        "168": [0, 0.69444, 0, 0, 0.55],
-        "176": [0, 0.69444, 0, 0, 0.73334],
-        "180": [0, 0.69444, 0, 0, 0.55],
-        "305": [0, 0.45833, 0, 0, 0.25556],
-        "567": [0.19444, 0.45833, 0, 0, 0.28611],
-        "710": [0, 0.69444, 0, 0, 0.55],
-        "711": [0, 0.63542, 0, 0, 0.55],
-        "713": [0, 0.63778, 0, 0, 0.55],
-        "728": [0, 0.69444, 0, 0, 0.55],
-        "729": [0, 0.69444, 0, 0, 0.30556],
-        "730": [0, 0.69444, 0, 0, 0.73334],
-        "732": [0, 0.69444, 0, 0, 0.55],
-        "733": [0, 0.69444, 0, 0, 0.55],
-        "915": [0, 0.69444, 0, 0, 0.58056],
-        "916": [0, 0.69444, 0, 0, 0.91667],
-        "920": [0, 0.69444, 0, 0, 0.85556],
-        "923": [0, 0.69444, 0, 0, 0.67223],
-        "926": [0, 0.69444, 0, 0, 0.73334],
-        "928": [0, 0.69444, 0, 0, 0.79445],
-        "931": [0, 0.69444, 0, 0, 0.79445],
-        "933": [0, 0.69444, 0, 0, 0.85556],
-        "934": [0, 0.69444, 0, 0, 0.79445],
-        "936": [0, 0.69444, 0, 0, 0.85556],
-        "937": [0, 0.69444, 0, 0, 0.79445],
-        "8211": [0, 0.45833, 0.03056, 0, 0.55],
-        "8212": [0, 0.45833, 0.03056, 0, 1.10001],
-        "8216": [0, 0.69444, 0, 0, 0.30556],
-        "8217": [0, 0.69444, 0, 0, 0.30556],
-        "8220": [0, 0.69444, 0, 0, 0.55834],
-        "8221": [0, 0.69444, 0, 0, 0.55834],
-    },
-    "SansSerif-Italic": {
-        "33": [0, 0.69444, 0.05733, 0, 0.31945],
-        "34": [0, 0.69444, 0.00316, 0, 0.5],
-        "35": [0.19444, 0.69444, 0.05087, 0, 0.83334],
-        "36": [0.05556, 0.75, 0.11156, 0, 0.5],
-        "37": [0.05556, 0.75, 0.03126, 0, 0.83334],
-        "38": [0, 0.69444, 0.03058, 0, 0.75834],
-        "39": [0, 0.69444, 0.07816, 0, 0.27778],
-        "40": [0.25, 0.75, 0.13164, 0, 0.38889],
-        "41": [0.25, 0.75, 0.02536, 0, 0.38889],
-        "42": [0, 0.75, 0.11775, 0, 0.5],
-        "43": [0.08333, 0.58333, 0.02536, 0, 0.77778],
-        "44": [0.125, 0.08333, 0, 0, 0.27778],
-        "45": [0, 0.44444, 0.01946, 0, 0.33333],
-        "46": [0, 0.08333, 0, 0, 0.27778],
-        "47": [0.25, 0.75, 0.13164, 0, 0.5],
-        "48": [0, 0.65556, 0.11156, 0, 0.5],
-        "49": [0, 0.65556, 0.11156, 0, 0.5],
-        "50": [0, 0.65556, 0.11156, 0, 0.5],
-        "51": [0, 0.65556, 0.11156, 0, 0.5],
-        "52": [0, 0.65556, 0.11156, 0, 0.5],
-        "53": [0, 0.65556, 0.11156, 0, 0.5],
-        "54": [0, 0.65556, 0.11156, 0, 0.5],
-        "55": [0, 0.65556, 0.11156, 0, 0.5],
-        "56": [0, 0.65556, 0.11156, 0, 0.5],
-        "57": [0, 0.65556, 0.11156, 0, 0.5],
-        "58": [0, 0.44444, 0.02502, 0, 0.27778],
-        "59": [0.125, 0.44444, 0.02502, 0, 0.27778],
-        "61": [-0.13, 0.37, 0.05087, 0, 0.77778],
-        "63": [0, 0.69444, 0.11809, 0, 0.47222],
-        "64": [0, 0.69444, 0.07555, 0, 0.66667],
-        "65": [0, 0.69444, 0, 0, 0.66667],
-        "66": [0, 0.69444, 0.08293, 0, 0.66667],
-        "67": [0, 0.69444, 0.11983, 0, 0.63889],
-        "68": [0, 0.69444, 0.07555, 0, 0.72223],
-        "69": [0, 0.69444, 0.11983, 0, 0.59722],
-        "70": [0, 0.69444, 0.13372, 0, 0.56945],
-        "71": [0, 0.69444, 0.11983, 0, 0.66667],
-        "72": [0, 0.69444, 0.08094, 0, 0.70834],
-        "73": [0, 0.69444, 0.13372, 0, 0.27778],
-        "74": [0, 0.69444, 0.08094, 0, 0.47222],
-        "75": [0, 0.69444, 0.11983, 0, 0.69445],
-        "76": [0, 0.69444, 0, 0, 0.54167],
-        "77": [0, 0.69444, 0.08094, 0, 0.875],
-        "78": [0, 0.69444, 0.08094, 0, 0.70834],
-        "79": [0, 0.69444, 0.07555, 0, 0.73611],
-        "80": [0, 0.69444, 0.08293, 0, 0.63889],
-        "81": [0.125, 0.69444, 0.07555, 0, 0.73611],
-        "82": [0, 0.69444, 0.08293, 0, 0.64584],
-        "83": [0, 0.69444, 0.09205, 0, 0.55556],
-        "84": [0, 0.69444, 0.13372, 0, 0.68056],
-        "85": [0, 0.69444, 0.08094, 0, 0.6875],
-        "86": [0, 0.69444, 0.1615, 0, 0.66667],
-        "87": [0, 0.69444, 0.1615, 0, 0.94445],
-        "88": [0, 0.69444, 0.13372, 0, 0.66667],
-        "89": [0, 0.69444, 0.17261, 0, 0.66667],
-        "90": [0, 0.69444, 0.11983, 0, 0.61111],
-        "91": [0.25, 0.75, 0.15942, 0, 0.28889],
-        "93": [0.25, 0.75, 0.08719, 0, 0.28889],
-        "94": [0, 0.69444, 0.0799, 0, 0.5],
-        "95": [0.35, 0.09444, 0.08616, 0, 0.5],
-        "97": [0, 0.44444, 0.00981, 0, 0.48056],
-        "98": [0, 0.69444, 0.03057, 0, 0.51667],
-        "99": [0, 0.44444, 0.08336, 0, 0.44445],
-        "100": [0, 0.69444, 0.09483, 0, 0.51667],
-        "101": [0, 0.44444, 0.06778, 0, 0.44445],
-        "102": [0, 0.69444, 0.21705, 0, 0.30556],
-        "103": [0.19444, 0.44444, 0.10836, 0, 0.5],
-        "104": [0, 0.69444, 0.01778, 0, 0.51667],
-        "105": [0, 0.67937, 0.09718, 0, 0.23889],
-        "106": [0.19444, 0.67937, 0.09162, 0, 0.26667],
-        "107": [0, 0.69444, 0.08336, 0, 0.48889],
-        "108": [0, 0.69444, 0.09483, 0, 0.23889],
-        "109": [0, 0.44444, 0.01778, 0, 0.79445],
-        "110": [0, 0.44444, 0.01778, 0, 0.51667],
-        "111": [0, 0.44444, 0.06613, 0, 0.5],
-        "112": [0.19444, 0.44444, 0.0389, 0, 0.51667],
-        "113": [0.19444, 0.44444, 0.04169, 0, 0.51667],
-        "114": [0, 0.44444, 0.10836, 0, 0.34167],
-        "115": [0, 0.44444, 0.0778, 0, 0.38333],
-        "116": [0, 0.57143, 0.07225, 0, 0.36111],
-        "117": [0, 0.44444, 0.04169, 0, 0.51667],
-        "118": [0, 0.44444, 0.10836, 0, 0.46111],
-        "119": [0, 0.44444, 0.10836, 0, 0.68334],
-        "120": [0, 0.44444, 0.09169, 0, 0.46111],
-        "121": [0.19444, 0.44444, 0.10836, 0, 0.46111],
-        "122": [0, 0.44444, 0.08752, 0, 0.43472],
-        "126": [0.35, 0.32659, 0.08826, 0, 0.5],
-        "168": [0, 0.67937, 0.06385, 0, 0.5],
-        "176": [0, 0.69444, 0, 0, 0.73752],
-        "305": [0, 0.44444, 0.04169, 0, 0.23889],
-        "567": [0.19444, 0.44444, 0.04169, 0, 0.26667],
-        "710": [0, 0.69444, 0.0799, 0, 0.5],
-        "711": [0, 0.63194, 0.08432, 0, 0.5],
-        "713": [0, 0.60889, 0.08776, 0, 0.5],
-        "714": [0, 0.69444, 0.09205, 0, 0.5],
-        "715": [0, 0.69444, 0, 0, 0.5],
-        "728": [0, 0.69444, 0.09483, 0, 0.5],
-        "729": [0, 0.67937, 0.07774, 0, 0.27778],
-        "730": [0, 0.69444, 0, 0, 0.73752],
-        "732": [0, 0.67659, 0.08826, 0, 0.5],
-        "733": [0, 0.69444, 0.09205, 0, 0.5],
-        "915": [0, 0.69444, 0.13372, 0, 0.54167],
-        "916": [0, 0.69444, 0, 0, 0.83334],
-        "920": [0, 0.69444, 0.07555, 0, 0.77778],
-        "923": [0, 0.69444, 0, 0, 0.61111],
-        "926": [0, 0.69444, 0.12816, 0, 0.66667],
-        "928": [0, 0.69444, 0.08094, 0, 0.70834],
-        "931": [0, 0.69444, 0.11983, 0, 0.72222],
-        "933": [0, 0.69444, 0.09031, 0, 0.77778],
-        "934": [0, 0.69444, 0.04603, 0, 0.72222],
-        "936": [0, 0.69444, 0.09031, 0, 0.77778],
-        "937": [0, 0.69444, 0.08293, 0, 0.72222],
-        "8211": [0, 0.44444, 0.08616, 0, 0.5],
-        "8212": [0, 0.44444, 0.08616, 0, 1.0],
-        "8216": [0, 0.69444, 0.07816, 0, 0.27778],
-        "8217": [0, 0.69444, 0.07816, 0, 0.27778],
-        "8220": [0, 0.69444, 0.14205, 0, 0.5],
-        "8221": [0, 0.69444, 0.00316, 0, 0.5],
-    },
-    "SansSerif-Regular": {
-        "33": [0, 0.69444, 0, 0, 0.31945],
-        "34": [0, 0.69444, 0, 0, 0.5],
-        "35": [0.19444, 0.69444, 0, 0, 0.83334],
-        "36": [0.05556, 0.75, 0, 0, 0.5],
-        "37": [0.05556, 0.75, 0, 0, 0.83334],
-        "38": [0, 0.69444, 0, 0, 0.75834],
-        "39": [0, 0.69444, 0, 0, 0.27778],
-        "40": [0.25, 0.75, 0, 0, 0.38889],
-        "41": [0.25, 0.75, 0, 0, 0.38889],
-        "42": [0, 0.75, 0, 0, 0.5],
-        "43": [0.08333, 0.58333, 0, 0, 0.77778],
-        "44": [0.125, 0.08333, 0, 0, 0.27778],
-        "45": [0, 0.44444, 0, 0, 0.33333],
-        "46": [0, 0.08333, 0, 0, 0.27778],
-        "47": [0.25, 0.75, 0, 0, 0.5],
-        "48": [0, 0.65556, 0, 0, 0.5],
-        "49": [0, 0.65556, 0, 0, 0.5],
-        "50": [0, 0.65556, 0, 0, 0.5],
-        "51": [0, 0.65556, 0, 0, 0.5],
-        "52": [0, 0.65556, 0, 0, 0.5],
-        "53": [0, 0.65556, 0, 0, 0.5],
-        "54": [0, 0.65556, 0, 0, 0.5],
-        "55": [0, 0.65556, 0, 0, 0.5],
-        "56": [0, 0.65556, 0, 0, 0.5],
-        "57": [0, 0.65556, 0, 0, 0.5],
-        "58": [0, 0.44444, 0, 0, 0.27778],
-        "59": [0.125, 0.44444, 0, 0, 0.27778],
-        "61": [-0.13, 0.37, 0, 0, 0.77778],
-        "63": [0, 0.69444, 0, 0, 0.47222],
-        "64": [0, 0.69444, 0, 0, 0.66667],
-        "65": [0, 0.69444, 0, 0, 0.66667],
-        "66": [0, 0.69444, 0, 0, 0.66667],
-        "67": [0, 0.69444, 0, 0, 0.63889],
-        "68": [0, 0.69444, 0, 0, 0.72223],
-        "69": [0, 0.69444, 0, 0, 0.59722],
-        "70": [0, 0.69444, 0, 0, 0.56945],
-        "71": [0, 0.69444, 0, 0, 0.66667],
-        "72": [0, 0.69444, 0, 0, 0.70834],
-        "73": [0, 0.69444, 0, 0, 0.27778],
-        "74": [0, 0.69444, 0, 0, 0.47222],
-        "75": [0, 0.69444, 0, 0, 0.69445],
-        "76": [0, 0.69444, 0, 0, 0.54167],
-        "77": [0, 0.69444, 0, 0, 0.875],
-        "78": [0, 0.69444, 0, 0, 0.70834],
-        "79": [0, 0.69444, 0, 0, 0.73611],
-        "80": [0, 0.69444, 0, 0, 0.63889],
-        "81": [0.125, 0.69444, 0, 0, 0.73611],
-        "82": [0, 0.69444, 0, 0, 0.64584],
-        "83": [0, 0.69444, 0, 0, 0.55556],
-        "84": [0, 0.69444, 0, 0, 0.68056],
-        "85": [0, 0.69444, 0, 0, 0.6875],
-        "86": [0, 0.69444, 0.01389, 0, 0.66667],
-        "87": [0, 0.69444, 0.01389, 0, 0.94445],
-        "88": [0, 0.69444, 0, 0, 0.66667],
-        "89": [0, 0.69444, 0.025, 0, 0.66667],
-        "90": [0, 0.69444, 0, 0, 0.61111],
-        "91": [0.25, 0.75, 0, 0, 0.28889],
-        "93": [0.25, 0.75, 0, 0, 0.28889],
-        "94": [0, 0.69444, 0, 0, 0.5],
-        "95": [0.35, 0.09444, 0.02778, 0, 0.5],
-        "97": [0, 0.44444, 0, 0, 0.48056],
-        "98": [0, 0.69444, 0, 0, 0.51667],
-        "99": [0, 0.44444, 0, 0, 0.44445],
-        "100": [0, 0.69444, 0, 0, 0.51667],
-        "101": [0, 0.44444, 0, 0, 0.44445],
-        "102": [0, 0.69444, 0.06944, 0, 0.30556],
-        "103": [0.19444, 0.44444, 0.01389, 0, 0.5],
-        "104": [0, 0.69444, 0, 0, 0.51667],
-        "105": [0, 0.67937, 0, 0, 0.23889],
-        "106": [0.19444, 0.67937, 0, 0, 0.26667],
-        "107": [0, 0.69444, 0, 0, 0.48889],
-        "108": [0, 0.69444, 0, 0, 0.23889],
-        "109": [0, 0.44444, 0, 0, 0.79445],
-        "110": [0, 0.44444, 0, 0, 0.51667],
-        "111": [0, 0.44444, 0, 0, 0.5],
-        "112": [0.19444, 0.44444, 0, 0, 0.51667],
-        "113": [0.19444, 0.44444, 0, 0, 0.51667],
-        "114": [0, 0.44444, 0.01389, 0, 0.34167],
-        "115": [0, 0.44444, 0, 0, 0.38333],
-        "116": [0, 0.57143, 0, 0, 0.36111],
-        "117": [0, 0.44444, 0, 0, 0.51667],
-        "118": [0, 0.44444, 0.01389, 0, 0.46111],
-        "119": [0, 0.44444, 0.01389, 0, 0.68334],
-        "120": [0, 0.44444, 0, 0, 0.46111],
-        "121": [0.19444, 0.44444, 0.01389, 0, 0.46111],
-        "122": [0, 0.44444, 0, 0, 0.43472],
-        "126": [0.35, 0.32659, 0, 0, 0.5],
-        "176": [0, 0.69444, 0, 0, 0.66667],
-        "305": [0, 0.44444, 0, 0, 0.23889],
-        "567": [0.19444, 0.44444, 0, 0, 0.26667],
-        "710": [0, 0.69444, 0, 0, 0.5],
-        "711": [0, 0.63194, 0, 0, 0.5],
-        "713": [0, 0.60889, 0, 0, 0.5],
-        "714": [0, 0.69444, 0, 0, 0.5],
-        "728": [0, 0.69444, 0, 0, 0.5],
-        "729": [0, 0.67937, 0, 0, 0.27778],
-        "730": [0, 0.69444, 0, 0, 0.66667],
-        "733": [0, 0.69444, 0, 0, 0.5],
-        "771": [0, 0.67659, 0, 0, 0.5],
-        "776": [0, 0.67937, 0, 0, 0.5],
-        "915": [0, 0.69444, 0, 0, 0.54167],
-        "916": [0, 0.69444, 0, 0, 0.83334],
-        "920": [0, 0.69444, 0, 0, 0.77778],
-        "923": [0, 0.69444, 0, 0, 0.61111],
-        "926": [0, 0.69444, 0, 0, 0.66667],
-        "928": [0, 0.69444, 0, 0, 0.70834],
-        "931": [0, 0.69444, 0, 0, 0.72222],
-        "933": [0, 0.69444, 0, 0, 0.77778],
-        "934": [0, 0.69444, 0, 0, 0.72222],
-        "936": [0, 0.69444, 0, 0, 0.77778],
-        "937": [0, 0.69444, 0, 0, 0.72222],
-        "8211": [0, 0.44444, 0.02778, 0, 0.5],
-        "8212": [0, 0.44444, 0.02778, 0, 1.0],
-        "8216": [0, 0.69444, 0, 0, 0.27778],
-        "8217": [0, 0.69444, 0, 0, 0.27778],
-        "8220": [0, 0.69444, 0, 0, 0.5],
-        "8221": [0, 0.69444, 0, 0, 0.5],
-    },
-    "Script-Regular": {
-        "65": [0, 0.7, 0.22925, 0, 0.80253],
-        "66": [0, 0.7, 0.04087, 0, 0.90757],
-        "67": [0, 0.7, 0.1689, 0, 0.66619],
-        "68": [0, 0.7, 0.09371, 0, 0.77443],
-        "69": [0, 0.7, 0.18583, 0, 0.56162],
-        "70": [0, 0.7, 0.13634, 0, 0.89544],
-        "71": [0, 0.7, 0.17322, 0, 0.60961],
-        "72": [0, 0.7, 0.29694, 0, 0.96919],
-        "73": [0, 0.7, 0.19189, 0, 0.80907],
-        "74": [0.27778, 0.7, 0.19189, 0, 1.05159],
-        "75": [0, 0.7, 0.31259, 0, 0.91364],
-        "76": [0, 0.7, 0.19189, 0, 0.87373],
-        "77": [0, 0.7, 0.15981, 0, 1.08031],
-        "78": [0, 0.7, 0.3525, 0, 0.9015],
-        "79": [0, 0.7, 0.08078, 0, 0.73787],
-        "80": [0, 0.7, 0.08078, 0, 1.01262],
-        "81": [0, 0.7, 0.03305, 0, 0.88282],
-        "82": [0, 0.7, 0.06259, 0, 0.85],
-        "83": [0, 0.7, 0.19189, 0, 0.86767],
-        "84": [0, 0.7, 0.29087, 0, 0.74697],
-        "85": [0, 0.7, 0.25815, 0, 0.79996],
-        "86": [0, 0.7, 0.27523, 0, 0.62204],
-        "87": [0, 0.7, 0.27523, 0, 0.80532],
-        "88": [0, 0.7, 0.26006, 0, 0.94445],
-        "89": [0, 0.7, 0.2939, 0, 0.70961],
-        "90": [0, 0.7, 0.24037, 0, 0.8212],
-    },
-    "Size1-Regular": {
-        "40": [0.35001, 0.85, 0, 0, 0.45834],
-        "41": [0.35001, 0.85, 0, 0, 0.45834],
-        "47": [0.35001, 0.85, 0, 0, 0.57778],
-        "91": [0.35001, 0.85, 0, 0, 0.41667],
-        "92": [0.35001, 0.85, 0, 0, 0.57778],
-        "93": [0.35001, 0.85, 0, 0, 0.41667],
-        "123": [0.35001, 0.85, 0, 0, 0.58334],
-        "125": [0.35001, 0.85, 0, 0, 0.58334],
-        "710": [0, 0.72222, 0, 0, 0.55556],
-        "732": [0, 0.72222, 0, 0, 0.55556],
-        "770": [0, 0.72222, 0, 0, 0.55556],
-        "771": [0, 0.72222, 0, 0, 0.55556],
-        "8214": [-0.00099, 0.601, 0, 0, 0.77778],
-        "8593": [1e-05, 0.6, 0, 0, 0.66667],
-        "8595": [1e-05, 0.6, 0, 0, 0.66667],
-        "8657": [1e-05, 0.6, 0, 0, 0.77778],
-        "8659": [1e-05, 0.6, 0, 0, 0.77778],
-        "8719": [0.25001, 0.75, 0, 0, 0.94445],
-        "8720": [0.25001, 0.75, 0, 0, 0.94445],
-        "8721": [0.25001, 0.75, 0, 0, 1.05556],
-        "8730": [0.35001, 0.85, 0, 0, 1.0],
-        "8739": [-0.00599, 0.606, 0, 0, 0.33333],
-        "8741": [-0.00599, 0.606, 0, 0, 0.55556],
-        "8747": [0.30612, 0.805, 0.19445, 0, 0.47222],
-        "8748": [0.306, 0.805, 0.19445, 0, 0.47222],
-        "8749": [0.306, 0.805, 0.19445, 0, 0.47222],
-        "8750": [0.30612, 0.805, 0.19445, 0, 0.47222],
-        "8896": [0.25001, 0.75, 0, 0, 0.83334],
-        "8897": [0.25001, 0.75, 0, 0, 0.83334],
-        "8898": [0.25001, 0.75, 0, 0, 0.83334],
-        "8899": [0.25001, 0.75, 0, 0, 0.83334],
-        "8968": [0.35001, 0.85, 0, 0, 0.47222],
-        "8969": [0.35001, 0.85, 0, 0, 0.47222],
-        "8970": [0.35001, 0.85, 0, 0, 0.47222],
-        "8971": [0.35001, 0.85, 0, 0, 0.47222],
-        "9168": [-0.00099, 0.601, 0, 0, 0.66667],
-        "10216": [0.35001, 0.85, 0, 0, 0.47222],
-        "10217": [0.35001, 0.85, 0, 0, 0.47222],
-        "10752": [0.25001, 0.75, 0, 0, 1.11111],
-        "10753": [0.25001, 0.75, 0, 0, 1.11111],
-        "10754": [0.25001, 0.75, 0, 0, 1.11111],
-        "10756": [0.25001, 0.75, 0, 0, 0.83334],
-        "10758": [0.25001, 0.75, 0, 0, 0.83334],
-    },
-    "Size2-Regular": {
-        "40": [0.65002, 1.15, 0, 0, 0.59722],
-        "41": [0.65002, 1.15, 0, 0, 0.59722],
-        "47": [0.65002, 1.15, 0, 0, 0.81111],
-        "91": [0.65002, 1.15, 0, 0, 0.47222],
-        "92": [0.65002, 1.15, 0, 0, 0.81111],
-        "93": [0.65002, 1.15, 0, 0, 0.47222],
-        "123": [0.65002, 1.15, 0, 0, 0.66667],
-        "125": [0.65002, 1.15, 0, 0, 0.66667],
-        "710": [0, 0.75, 0, 0, 1.0],
-        "732": [0, 0.75, 0, 0, 1.0],
-        "770": [0, 0.75, 0, 0, 1.0],
-        "771": [0, 0.75, 0, 0, 1.0],
-        "8719": [0.55001, 1.05, 0, 0, 1.27778],
-        "8720": [0.55001, 1.05, 0, 0, 1.27778],
-        "8721": [0.55001, 1.05, 0, 0, 1.44445],
-        "8730": [0.65002, 1.15, 0, 0, 1.0],
-        "8747": [0.86225, 1.36, 0.44445, 0, 0.55556],
-        "8748": [0.862, 1.36, 0.44445, 0, 0.55556],
-        "8749": [0.862, 1.36, 0.44445, 0, 0.55556],
-        "8750": [0.86225, 1.36, 0.44445, 0, 0.55556],
-        "8896": [0.55001, 1.05, 0, 0, 1.11111],
-        "8897": [0.55001, 1.05, 0, 0, 1.11111],
-        "8898": [0.55001, 1.05, 0, 0, 1.11111],
-        "8899": [0.55001, 1.05, 0, 0, 1.11111],
-        "8968": [0.65002, 1.15, 0, 0, 0.52778],
-        "8969": [0.65002, 1.15, 0, 0, 0.52778],
-        "8970": [0.65002, 1.15, 0, 0, 0.52778],
-        "8971": [0.65002, 1.15, 0, 0, 0.52778],
-        "10216": [0.65002, 1.15, 0, 0, 0.61111],
-        "10217": [0.65002, 1.15, 0, 0, 0.61111],
-        "10752": [0.55001, 1.05, 0, 0, 1.51112],
-        "10753": [0.55001, 1.05, 0, 0, 1.51112],
-        "10754": [0.55001, 1.05, 0, 0, 1.51112],
-        "10756": [0.55001, 1.05, 0, 0, 1.11111],
-        "10758": [0.55001, 1.05, 0, 0, 1.11111],
-    },
-    "Size3-Regular": {
-        "40": [0.95003, 1.45, 0, 0, 0.73611],
-        "41": [0.95003, 1.45, 0, 0, 0.73611],
-        "47": [0.95003, 1.45, 0, 0, 1.04445],
-        "91": [0.95003, 1.45, 0, 0, 0.52778],
-        "92": [0.95003, 1.45, 0, 0, 1.04445],
-        "93": [0.95003, 1.45, 0, 0, 0.52778],
-        "123": [0.95003, 1.45, 0, 0, 0.75],
-        "125": [0.95003, 1.45, 0, 0, 0.75],
-        "710": [0, 0.75, 0, 0, 1.44445],
-        "732": [0, 0.75, 0, 0, 1.44445],
-        "770": [0, 0.75, 0, 0, 1.44445],
-        "771": [0, 0.75, 0, 0, 1.44445],
-        "8730": [0.95003, 1.45, 0, 0, 1.0],
-        "8968": [0.95003, 1.45, 0, 0, 0.58334],
-        "8969": [0.95003, 1.45, 0, 0, 0.58334],
-        "8970": [0.95003, 1.45, 0, 0, 0.58334],
-        "8971": [0.95003, 1.45, 0, 0, 0.58334],
-        "10216": [0.95003, 1.45, 0, 0, 0.75],
-        "10217": [0.95003, 1.45, 0, 0, 0.75],
-    },
-    "Size4-Regular": {
-        "40": [1.25003, 1.75, 0, 0, 0.79167],
-        "41": [1.25003, 1.75, 0, 0, 0.79167],
-        "47": [1.25003, 1.75, 0, 0, 1.27778],
-        "91": [1.25003, 1.75, 0, 0, 0.58334],
-        "92": [1.25003, 1.75, 0, 0, 1.27778],
-        "93": [1.25003, 1.75, 0, 0, 0.58334],
-        "123": [1.25003, 1.75, 0, 0, 0.80556],
-        "125": [1.25003, 1.75, 0, 0, 0.80556],
-        "710": [0, 0.825, 0, 0, 1.8889],
-        "732": [0, 0.825, 0, 0, 1.8889],
-        "770": [0, 0.825, 0, 0, 1.8889],
-        "771": [0, 0.825, 0, 0, 1.8889],
-        "8730": [1.25003, 1.75, 0, 0, 1.0],
-        "8968": [1.25003, 1.75, 0, 0, 0.63889],
-        "8969": [1.25003, 1.75, 0, 0, 0.63889],
-        "8970": [1.25003, 1.75, 0, 0, 0.63889],
-        "8971": [1.25003, 1.75, 0, 0, 0.63889],
-        "9115": [0.64502, 1.155, 0, 0, 0.875],
-        "9116": [1e-05, 0.6, 0, 0, 0.875],
-        "9117": [0.64502, 1.155, 0, 0, 0.875],
-        "9118": [0.64502, 1.155, 0, 0, 0.875],
-        "9119": [1e-05, 0.6, 0, 0, 0.875],
-        "9120": [0.64502, 1.155, 0, 0, 0.875],
-        "9121": [0.64502, 1.155, 0, 0, 0.66667],
-        "9122": [-0.00099, 0.601, 0, 0, 0.66667],
-        "9123": [0.64502, 1.155, 0, 0, 0.66667],
-        "9124": [0.64502, 1.155, 0, 0, 0.66667],
-        "9125": [-0.00099, 0.601, 0, 0, 0.66667],
-        "9126": [0.64502, 1.155, 0, 0, 0.66667],
-        "9127": [1e-05, 0.9, 0, 0, 0.88889],
-        "9128": [0.65002, 1.15, 0, 0, 0.88889],
-        "9129": [0.90001, 0, 0, 0, 0.88889],
-        "9130": [0, 0.3, 0, 0, 0.88889],
-        "9131": [1e-05, 0.9, 0, 0, 0.88889],
-        "9132": [0.65002, 1.15, 0, 0, 0.88889],
-        "9133": [0.90001, 0, 0, 0, 0.88889],
-        "9143": [0.88502, 0.915, 0, 0, 1.05556],
-        "10216": [1.25003, 1.75, 0, 0, 0.80556],
-        "10217": [1.25003, 1.75, 0, 0, 0.80556],
-        "57344": [-0.00499, 0.605, 0, 0, 1.05556],
-        "57345": [-0.00499, 0.605, 0, 0, 1.05556],
-        "57680": [0, 0.12, 0, 0, 0.45],
-        "57681": [0, 0.12, 0, 0, 0.45],
-        "57682": [0, 0.12, 0, 0, 0.45],
-        "57683": [0, 0.12, 0, 0, 0.45],
-    },
-    "Typewriter-Regular": {
-        "33": [0, 0.61111, 0, 0, 0.525],
-        "34": [0, 0.61111, 0, 0, 0.525],
-        "35": [0, 0.61111, 0, 0, 0.525],
-        "36": [0.08333, 0.69444, 0, 0, 0.525],
-        "37": [0.08333, 0.69444, 0, 0, 0.525],
-        "38": [0, 0.61111, 0, 0, 0.525],
-        "39": [0, 0.61111, 0, 0, 0.525],
-        "40": [0.08333, 0.69444, 0, 0, 0.525],
-        "41": [0.08333, 0.69444, 0, 0, 0.525],
-        "42": [0, 0.52083, 0, 0, 0.525],
-        "43": [-0.08056, 0.53055, 0, 0, 0.525],
-        "44": [0.13889, 0.125, 0, 0, 0.525],
-        "45": [-0.08056, 0.53055, 0, 0, 0.525],
-        "46": [0, 0.125, 0, 0, 0.525],
-        "47": [0.08333, 0.69444, 0, 0, 0.525],
-        "48": [0, 0.61111, 0, 0, 0.525],
-        "49": [0, 0.61111, 0, 0, 0.525],
-        "50": [0, 0.61111, 0, 0, 0.525],
-        "51": [0, 0.61111, 0, 0, 0.525],
-        "52": [0, 0.61111, 0, 0, 0.525],
-        "53": [0, 0.61111, 0, 0, 0.525],
-        "54": [0, 0.61111, 0, 0, 0.525],
-        "55": [0, 0.61111, 0, 0, 0.525],
-        "56": [0, 0.61111, 0, 0, 0.525],
-        "57": [0, 0.61111, 0, 0, 0.525],
-        "58": [0, 0.43056, 0, 0, 0.525],
-        "59": [0.13889, 0.43056, 0, 0, 0.525],
-        "60": [-0.05556, 0.55556, 0, 0, 0.525],
-        "61": [-0.19549, 0.41562, 0, 0, 0.525],
-        "62": [-0.05556, 0.55556, 0, 0, 0.525],
-        "63": [0, 0.61111, 0, 0, 0.525],
-        "64": [0, 0.61111, 0, 0, 0.525],
-        "65": [0, 0.61111, 0, 0, 0.525],
-        "66": [0, 0.61111, 0, 0, 0.525],
-        "67": [0, 0.61111, 0, 0, 0.525],
-        "68": [0, 0.61111, 0, 0, 0.525],
-        "69": [0, 0.61111, 0, 0, 0.525],
-        "70": [0, 0.61111, 0, 0, 0.525],
-        "71": [0, 0.61111, 0, 0, 0.525],
-        "72": [0, 0.61111, 0, 0, 0.525],
-        "73": [0, 0.61111, 0, 0, 0.525],
-        "74": [0, 0.61111, 0, 0, 0.525],
-        "75": [0, 0.61111, 0, 0, 0.525],
-        "76": [0, 0.61111, 0, 0, 0.525],
-        "77": [0, 0.61111, 0, 0, 0.525],
-        "78": [0, 0.61111, 0, 0, 0.525],
-        "79": [0, 0.61111, 0, 0, 0.525],
-        "80": [0, 0.61111, 0, 0, 0.525],
-        "81": [0.13889, 0.61111, 0, 0, 0.525],
-        "82": [0, 0.61111, 0, 0, 0.525],
-        "83": [0, 0.61111, 0, 0, 0.525],
-        "84": [0, 0.61111, 0, 0, 0.525],
-        "85": [0, 0.61111, 0, 0, 0.525],
-        "86": [0, 0.61111, 0, 0, 0.525],
-        "87": [0, 0.61111, 0, 0, 0.525],
-        "88": [0, 0.61111, 0, 0, 0.525],
-        "89": [0, 0.61111, 0, 0, 0.525],
-        "90": [0, 0.61111, 0, 0, 0.525],
-        "91": [0.08333, 0.69444, 0, 0, 0.525],
-        "92": [0.08333, 0.69444, 0, 0, 0.525],
-        "93": [0.08333, 0.69444, 0, 0, 0.525],
-        "94": [0, 0.61111, 0, 0, 0.525],
-        "95": [0.09514, 0, 0, 0, 0.525],
-        "96": [0, 0.61111, 0, 0, 0.525],
-        "97": [0, 0.43056, 0, 0, 0.525],
-        "98": [0, 0.61111, 0, 0, 0.525],
-        "99": [0, 0.43056, 0, 0, 0.525],
-        "100": [0, 0.61111, 0, 0, 0.525],
-        "101": [0, 0.43056, 0, 0, 0.525],
-        "102": [0, 0.61111, 0, 0, 0.525],
-        "103": [0.22222, 0.43056, 0, 0, 0.525],
-        "104": [0, 0.61111, 0, 0, 0.525],
-        "105": [0, 0.61111, 0, 0, 0.525],
-        "106": [0.22222, 0.61111, 0, 0, 0.525],
-        "107": [0, 0.61111, 0, 0, 0.525],
-        "108": [0, 0.61111, 0, 0, 0.525],
-        "109": [0, 0.43056, 0, 0, 0.525],
-        "110": [0, 0.43056, 0, 0, 0.525],
-        "111": [0, 0.43056, 0, 0, 0.525],
-        "112": [0.22222, 0.43056, 0, 0, 0.525],
-        "113": [0.22222, 0.43056, 0, 0, 0.525],
-        "114": [0, 0.43056, 0, 0, 0.525],
-        "115": [0, 0.43056, 0, 0, 0.525],
-        "116": [0, 0.55358, 0, 0, 0.525],
-        "117": [0, 0.43056, 0, 0, 0.525],
-        "118": [0, 0.43056, 0, 0, 0.525],
-        "119": [0, 0.43056, 0, 0, 0.525],
-        "120": [0, 0.43056, 0, 0, 0.525],
-        "121": [0.22222, 0.43056, 0, 0, 0.525],
-        "122": [0, 0.43056, 0, 0, 0.525],
-        "123": [0.08333, 0.69444, 0, 0, 0.525],
-        "124": [0.08333, 0.69444, 0, 0, 0.525],
-        "125": [0.08333, 0.69444, 0, 0, 0.525],
-        "126": [0, 0.61111, 0, 0, 0.525],
-        "127": [0, 0.61111, 0, 0, 0.525],
-        "176": [0, 0.61111, 0, 0, 0.525],
-        "305": [0, 0.43056, 0, 0, 0.525],
-        "567": [0.22222, 0.43056, 0, 0, 0.525],
-        "711": [0, 0.56597, 0, 0, 0.525],
-        "713": [0, 0.56555, 0, 0, 0.525],
-        "714": [0, 0.61111, 0, 0, 0.525],
-        "715": [0, 0.61111, 0, 0, 0.525],
-        "728": [0, 0.61111, 0, 0, 0.525],
-        "730": [0, 0.61111, 0, 0, 0.525],
-        "770": [0, 0.61111, 0, 0, 0.525],
-        "771": [0, 0.61111, 0, 0, 0.525],
-        "776": [0, 0.61111, 0, 0, 0.525],
-        "915": [0, 0.61111, 0, 0, 0.525],
-        "916": [0, 0.61111, 0, 0, 0.525],
-        "920": [0, 0.61111, 0, 0, 0.525],
-        "923": [0, 0.61111, 0, 0, 0.525],
-        "926": [0, 0.61111, 0, 0, 0.525],
-        "928": [0, 0.61111, 0, 0, 0.525],
-        "931": [0, 0.61111, 0, 0, 0.525],
-        "933": [0, 0.61111, 0, 0, 0.525],
-        "934": [0, 0.61111, 0, 0, 0.525],
-        "936": [0, 0.61111, 0, 0, 0.525],
-        "937": [0, 0.61111, 0, 0, 0.525],
-        "8216": [0, 0.61111, 0, 0, 0.525],
-        "8217": [0, 0.61111, 0, 0, 0.525],
-        "8242": [0, 0.61111, 0, 0, 0.525],
-        "9251": [0.11111, 0.21944, 0, 0, 0.525],
-    },
-};
